==> 1-merging.Rmd <==
---
title: "DHS data merging"
author: "Justin Guerra, Arnav Gupta, Andres Colubri"
output: html_document
date: "2024-07-31"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# libraries needed
library(tidyverse)  # most variable creation here uses tidyverse 
library(tidyselect) # used to select variables in FP_EVENTS.R
library(haven)      # used for Haven labeled DHSvariables
library(labelled)   # used for Haven labeled variable creation
library(expss)      # for creating tables with Haven labeled data
# library(xlsx)      # for exporting to excel
library(naniar)     # to use replace_with_na function
library(here)       # to get R project path
library(sjlabelled) # to set variables label
library(survey)     # to calculate weighted ratio for GAR
library(properties)
```

```{r}
prop <- read.properties("data.properties")
data_folder <- prop$data_folder
```

```{r}
# Household member df
HMdffile <- here(data_folder, "household-member", "IAPR7EDT", "IAPR7EFL.DTA")
HMdf <- read_dta(HMdffile)
HMdf
```

```{r}
# Children df
Cdffile <- here(data_folder, "children", "IAKR7EDT", "IAKR7EFL.DTA")
Cdf <- read_dta(Cdffile)
```

```{r}
# Rename variables in children df to then merge with household
Cdf <- Cdf %>% rename("hv000" = "v000")
Cdf <- Cdf %>% rename("hv001" = "v001")
Cdf <- Cdf %>% rename("hv002" = "v002")
Cdf <- Cdf %>% rename("hvidx" = "b16")
```

```{r}
#capture drop _merge
#gen b16 = hvidx
#gen v000 = hv000
#gen v001 = hv001
#gen v002 = hv002
#merge 1:1  v000 v001 v002 b16 using filename

df <- merge(HMdf, Cdf, by = c("hv000", "hv001", "hv002", "hvidx"))
df
```

```{r}
# Save merged data as csv
# csvfile <- here(data_folder, "merged.csv") 
# write.csv(df, csvfile, row.names=FALSE)
```

```{r}
# Save merged data as R object
rdafile <- here(data_folder, "merged.rda") 
save(df, file=rdafile) 
```
