==> 2-recoding.Rmd <==
---
title: "DHS data recoding"
author: "Justin Guerra, Arnav Gupta, Andres Colubri"
output: html_document
date: "2024-07-31"
---

From Stata to R
https://fsolt.org/blog/posts/switch-to-r/
https://epi-stats.github.io/Rtutorials/Stata_to_R#recode
https://www.techtips.surveydesign.com.au/post/the-recode-command


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# libraries needed
library(tidyverse)  # most variable creation here uses tidyverse 
library(tidyselect) # used to select variables in FP_EVENTS.R
library(haven)      # used for Haven labeled DHSvariables
library(labelled)   # used for Haven labeled variable creation
library(expss)      # for creating tables with Haven labeled df
library(readxl)     # for exporting to excel
library(naniar)     # to use replace_with_na function
library(here)       # to get R project path
library(sjlabelled) # to set variables label
library(survey)     # to calculate weighted ratio for GAR
library(Hmisc)
library(tidyr)
library(properties)
```

```{r}
prop <- read.properties("data.properties")
data_folder <- prop$data_folder
```

```{r}
datafile <- here(data_folder, "merged.rda") 
load(datafile)
```

**** Some utilities

```{r}
# function to convert to factor with order of levels using case_when:
# https://stackoverflow.com/questions/49572416/r-convert-to-factor-with-order-of-levels-same-with-case-when
fct_case_when <- function(...) {
  args <- as.list(match.call())
  levels <- sapply(args[-1], function(f) f[[3]])  # extract RHS of formula
  levels <- levels[!is.na(levels)]
  factor(dplyr::case_when(...), levels=levels)
}
```


**** List of recoded variables

```{r}
recoded_vars <- c()
```


**** Add
capture drop real_id
gen real_id =  real(string(hv001)+string(hv002)+string(hvidx))
cap drop house_id
gen house_id = real(string(hv001)+string(hv002))

```{r}
# check if `real_id` variable exists and drop if it does
if ("real_id" %in% names(df)) {
  df$real_id <- NULL
}

# create `real_id` variable as a concatenation of `hv001`, `hv002`, and `hvidx`
df$real_id <- as.numeric(paste0(df$hv001, df$hv002, df$hvidx))

# check if `house_id` variable exists and drop if it does
if ("house_id" %in% names(df)) {
  df$house_id <- NULL
}

# create `house_id` variable as a concatenation of `hv001` and `hv002`
df$house_id <- as.numeric(paste0(df$hv001, df$hv002))
```


capture drop wgt
generate wgt = hv005/1000000 
gen wgt2 = 1
*svyset [pw=wgt], psu(hv001) strata(hv022) singleunit(scaled)
svyset hv001, weight(wgt) strata(hv022) , singleunit(centered) || _n, weight(wgt2) 

```{r}
# Remove 'wgt' if it exists
if("wgt" %in% names(df)){
  df$wgt <- NULL
}

# Generate 'wgt' variable
df$wgt <- df$hv005 / 1000000

# Generate 'wgt2' variable
df$wgt2 <- 1

# Set up survey design (equivalent to the commented out Stata code)
# Uncomment the following lines if you want to use this survey design
# design <- svydesign(ids = ~hv001, strata = ~hv022, weights = ~wgt, data = df,
#                     nest = TRUE, singleunit = "scaled")

# Set up survey design with singleunit set to "centered"
design <- svydesign(
  ids = ~hv001, 
  strata = ~hv022, 
  weights = ~wgt, 
  data = df,
  nest = TRUE, 
  singleunit = "centered"
)

# Update the design with the wgt2 variable
design <- update(design, wgt2 = df$wgt2)

recoded_vars <- append(recoded_vars, "wgt2")
```


**** Identify eligible children ****
capture drop eligch 
gen eligch = 0 
replace eligch = 1 if (hv103==1 & hc70 < 9996 & hc71 < 9996 & hc72 < 9996)

```{r}

# Remove 'eligch' variable if it exists
if("eligch" %in% colnames(df)) {
  df$eligch <- NULL
}

# Generate eligch variable
df$eligch <- 0
df$eligch[df$hv103 == 1 & df$hc70 < 9996 & df$hc71 < 9996 & df$hc72 < 9996] <- 1

recoded_vars <- append(recoded_vars, "eligch")
```

capture drop defacto_child
gen defacto_child = hv103

capture drop real_id
gen real_id =  real(string(hv001)+string(hv002)+string(hvidx))

capture drop mother_id
gen mother_id =  real(string(hv001)+string(hv002)+string(hv112))
replace mother_id = . if midx == .

capture drop father_id
gen father_id =  real(string(hv001)+string(hv002)+string(hv114))
replace father_id = . if midx == .

```{r}
# Remove 'defacto_child' variable if it exists
if("defacto_child" %in% colnames(df)) {
  df$defacto_child <- NULL
}

# Generate defacto_child variable
df$defacto_child <- df$hv103

# Remove 'real_id' variable if it exists
if("real_id" %in% colnames(df)) {
  df$real_id <- NULL
}

# Generate real_id variable
df$real_id <- as.numeric(paste0(df$hv001, df$hv002, df$hvidx))

# Remove 'mother_id' variable if it exists
if("mother_id" %in% colnames(df)) {
  df$mother_id <- NULL
}

# Generate mother_id variable
df$mother_id <- as.numeric(paste0(df$hv001, df$hv002, df$hv112))
df$mother_id[is.na(df$midx)] <- NA

# Remove 'father_id' variable if it exists
if("father_id" %in% colnames(df)) {
  df$father_id <- NULL
}

# Generate father_id variable
df$father_id <- as.numeric(paste0(df$hv001, df$hv002, df$hv114))
df$father_id[is.na(df$midx)] <- NA

recoded_vars <- append(recoded_vars, "defacto_child")
```

**** Anthropometric variables ****
capture drop child_age 
gen child_age = hc1

capture drop age_group 
recode child_age (0/5 = 0 "0. 0-5 mo")(6/11 = 1 "1. 6-11 mo")(12/17 = 2 "2. 12-17 mo") ///
(18/24 = 3 "3. 18-24 mo")(25/59 = .),gen(age_group)
recode child_age (0/5 = 0 "0. 0-5 mo")(6/11 = 1 "1. 6-11 mo")(12/17 = 2 "2. 12-17 mo") ///
(18/23 = 3 "3. 18-23 mo") (24/35 = 4 "4. 24-35 mo")(36/47 = 5 "5. 36-47 mo")(48/59 = 6 "6. 48-60 mo"),gen(age_group_5y)


```{r}
if("child_age" %in% colnames(df)) {
  df$child_age <- NULL
}

# Generate child_age variable
df$child_age <- df$hc1

recoded_vars <- append(recoded_vars, "child_age")
```

```{r}
# Remove 'age_group' variable if it exists
if("age_group" %in% colnames(df)) {
  df$age_group <- NULL
}

# Generate age_group variable
df$age_group <- cut(df$child_age,
                           breaks = c(-Inf, 5, 11, 17, 24),
                           labels = c("0. 0-5 mo", "1. 6-11 mo", "2. 12-17 mo", "3. 18-24 mo"),
                           right = TRUE)

# Remove 'age_group_5y' variable if it exists
if("age_group_5y" %in% colnames(df)) {
  df$age_group_5y <- NULL
}

# Generate age_group_5y variable
df$age_group_5y <- cut(df$child_age,
                               breaks = c(-Inf, 5, 11, 17, 23, 35, 47, 60),
                               labels = c("0. 0-5 mo", "1. 6-11 mo", "2. 12-17 mo", "3. 18-23 mo", "4. 24-35 mo", "5. 36-47 mo", "6. 48-60 mo"),
                               right = TRUE)

recoded_vars <- append(recoded_vars, "age_group")
recoded_vars <- append(recoded_vars, "age_group_5y")
```


capture drop anthro_measured
gen anthro_measured = hc13

capture drop height_position 
gen height_position = hc15

capture drop haz waz whz bmiz
gen haz = hc70/100
replace haz = . if eligch == 0  
gen waz = hc71/100
replace waz = . if  eligch == 0   
gen whz = hc72/100
replace whz = . if eligch == 0 
gen bmiz = hc73/100
replace bmiz = . if eligch == 0 

capture drop stunted 
gen stunted = . 
replace stunted = 1 if haz < -2
replace stunted = 0 if haz >= -2 & haz != . 

capture drop o_stunted 
recode haz (0/6 = 0 "0. 0 or higher")(-2/-0.01 = 1 "1. < 0 and => -2  ") ///
(-3/-2.01 = 2 "2. < -2 and => -3  ")(-6/-3.01 = 3 "3. lower than -3"), gen(o_stunted)

```{r}
# Remove 'anthro_measured' variable if it exists
if("anthro_measured" %in% colnames(df)) {
  df$anthro_measured <- NULL
}

# Generate anthro_measured variable
df$anthro_measured <- df$hc13

# Remove 'height_position' variable if it exists
if("height_position" %in% colnames(df)) {
  df$height_position <- NULL
}

# Generate height_position variable
df$height_position <- df$hc15

recoded_vars <- append(recoded_vars, "anthro_measured")
recoded_vars <- append(recoded_vars, "height_position")
```

```{r}
# Remove haz, waz, whz, bmiz variables if they exist
df[, c("haz", "waz", "whz", "bmiz")] <- NULL

# Generate haz, waz, whz, bmiz variables
df$haz <- df$hc70 / 100
df$haz[df$eligch == 0] <- NA

df$waz <- df$hc71 / 100
df$waz[df$eligch == 0] <- NA

df$whz <- df$hc72 / 100
df$whz[df$eligch == 0] <- NA

df$bmiz <- df$hc73 / 100
df$bmiz[df$eligch == 0] <- NA

# Remove 'stunted' variable if it exists
if("stunted" %in% colnames(df)) {
  df$stunted <- NULL
}

# Generate stunted variable
df$stunted <- NA
df$stunted[df$haz < -2] <- 1
df$stunted[df$haz >= -2 & !is.na(df$haz)] <- 0

# Remove 'o_stunted' variable if it exists
if("o_stunted" %in% colnames(df)) {
  df$o_stunted <- NULL
}

# Generate o_stunted variable
df$o_stunted <- cut(df$haz,
                           breaks = c(-6, -3.01, -2.01, -0.01, Inf),
                           labels = c("3. lower than -3", "2. < -2 and => -3", "1. < 0 and => -2", "0. 0 or higher"),
                           right = FALSE)

recoded_vars <- append(recoded_vars, "haz") 
recoded_vars <- append(recoded_vars, "waz") 
recoded_vars <- append(recoded_vars, "whz")
recoded_vars <- append(recoded_vars, "bmiz")
recoded_vars <- append(recoded_vars, "stunted")
recoded_vars <- append(recoded_vars, "o_stunted")
```

capture drop underweight 
gen underweight = . 
replace underweight = 1 if waz < -2
replace underweight = 0 if waz >= -2 & waz != . 

```{r}
# Remove the "underweight" column if it exists
if ("underweight" %in% names(df)) {
  df$underweight <- NULL
}

# Create a new "underweight" column with missing values
df$underweight <- NA

# Replace missing values with 1 if waz < -2
df$underweight[df$waz < -2] <- 1

# Replace missing values with 0 if waz >= -2 and not missing
df$underweight[df$waz >= -2 & !is.na(df$waz)] <- 0

# Convert to factor
df$underweight <- as.factor(df$underweight)

recoded_vars <- append(recoded_vars, "underweight")
```

capture drop o_underweight
recode waz (0/6 = 0 "0. 0 or higher")(-2/-0.01 = 1 "1. < 0 and => -2  ") ///
(-3/-2.01 = 2 "2. < -2 and => -3  ")(-6/-3.01 = 3 "3. lower than -3"), gen(o_underweight)

```{r}
# Remove the "o_underweight" column if it exists
if ("o_underweight" %in% names(df)) {
  df$o_underweight <- NULL
}

# Recode waz into the new "o_underweight" column
df$o_underweight <- cut(df$waz, 
  breaks = c(-Inf, -3.01, -2.01, -0.01, 6), 
  labels = c("3. lower than -3", "2. < -2 and => -3", "1. < 0 and => -2", "0. 0 or higher"), 
  include.lowest = TRUE
)

recoded_vars <- append(recoded_vars, "o_underweight")
```

capture drop wasted 
gen wasted = . 
replace wasted = 1 if whz < -2
replace wasted = 0 if whz >= -2 & haz != . 
```{r}
# Remove the "wasted" column if it exists
if ("wasted" %in% names(df)) {
  df$wasted <- NULL
}

# Create a new "wasted" column with missing values
df$wasted <- NA

# Replace missing values with 1 if whz < -2
df$wasted[df$whz < -2] <- 1

# Replace missing values with 0 if whz >= -2 and not missing
df$wasted[df$whz >= -2 & !is.na(df$haz)] <- 0

#Convert to factor
df$wasted <- as.factor(df$wasted)

recoded_vars <- append(recoded_vars, "wasted")
```

capture drop o_wasted
recode whz (0/6 = 0 "0. 0 or higher")(-2/-0.01 = 1 "1. < 0 and => -2  ") ///
(-3/-2.01 = 2 "2. < -2 and => -3  ")(-6/-3.01 = 3 "3. lower than -3"), gen(o_wasted)

```{r}
# Remove the "o_wasted" column if it exists
if ("o_wasted" %in% names(df)) {
  df$o_wasted <- NULL
}

# Recode whz into the new "o_wasted" column
df$o_wasted <- cut(df$whz, 
  breaks = c(-Inf, -3.01, -2.01, -0.01, Inf), 
  labels = c("3. lower than -3", "2. < -2 and => -3", "1. < 0 and => -2", "0. 0 or higher"), 
  include.lowest = TRUE
)

recoded_vars <- append(recoded_vars, "o_wasted")
```

capture drop ciaf 
gen ciaf = 0 
replace ciaf = 1 if stunted == 1 | wasted == 1 | underweight == 1
replace ciaf = . if stunted == . | wasted == . & underweight == .

```{r}
# Remove the "ciaf" column if it exists
if ("ciaf" %in% names(df)) {
  df$ciaf <- NULL
}

# Create a new "ciaf" column with 0 values
df$ciaf <- 0

# Replace values with 1 if stunted, wasted, or underweight is 1
df$ciaf[df$stunted == 1 | df$wasted == 1 | df$underweight == 1] <- 1

# Replace missing values with NA
df$ciaf[is.na(df$stunted) | is.na(df$wasted) | is.na(df$underweight)] <- NA

recoded_vars <- append(recoded_vars, "ciaf")
```

capture drop stunt_waste
gen stunt_waste = 0
replace stunt_waste = 1 if stunted == 1 & wasted == 1
replace stunt_waste = . if stunted -- . & wasted == . 

```{r}
# Remove the "stunt_waste" column if it exists
if ("stunt_waste" %in% names(df)) {
  df$stunt_waste <- NULL
}

# Create a new "stunt_waste" column with 0 values
df$stunt_waste <- 0

# Replace values with 1 if both stunted and wasted are 1
df$stunt_waste[df$stunted == 1 & df$wasted == 1] <- 1

# Replace missing values with NA
df$stunt_waste[is.na(df$stunted) | is.na(df$wasted)] <- NA

# Convert to factor
df$stunt_waste <- as.factor(df$stunt_waste)

recoded_vars <- append(recoded_vars, "stunt_waste")
```


capture drop elig
gen elig = 0
replace elig = 1 if bidx == 1"

```{r}
# Remove the "elig" column if it exists
if ("elig" %in% names(df)) {
  df$elig <- NULL
}

# Create a new "elig" column with 0 values
df$elig <- 0

# Replace values with 1 if bidx is 1
df$elig[df$bidx == 1] <- 1

#Convert to factor
df$elig <- as.factor(df$elig)

recoded_vars <- append(recoded_vars, "elig")
```



*** Predictors ***

capture drop wealth_quintile 
gen wealth_quintile = v190 
label value wealth_quintile v190

```{r}
# Remove wealth_quintile if it exists in the df
if ("wealth_quintile" %in% colnames(df)) {
  df$wealth_quintile <- NULL
}

# Create a new variable wealth_quintile as a copy of v190
df <- df %>%
  mutate(wealth_quintile = v190 %>% as.factor())

recoded_vars <- append(recoded_vars, "wealth_quintile")
```


capture drop maternal_height
gen maternal_height_cm = v438/10 //height in centimeters
replace maternal_height_cm = . if maternal_height_cm > 999
recode maternal_height_cm (160/210 = 1 "1. 160+ cm") ///
(155/159.999 = 2 "2. 155-159.9 cm") (150/154.999 = 3 "3. 150-154.9 cm") ///
(145/149.999 = 4 "4. 145-149.9 cm") (80/144.999 = 5 "5. < 145 cm"), gen(maternal_height)

```{r}
# Remove variables starting with "maternal_height" if they exist
df <- df %>%
  select(-starts_with("maternal_height"))

# Create a new variable maternal_height_cm and recode it into maternal_height
df <- df %>%
  mutate(maternal_height_cm = v438 / 10) %>%
  mutate(maternal_height_cm = ifelse(maternal_height_cm > 999, NA, maternal_height_cm)) %>%
  mutate(maternal_height = fct_case_when(
    maternal_height_cm >= 160 & maternal_height_cm <= 210   ~ "1. 160+ cm",
    maternal_height_cm >= 155 & maternal_height_cm < 160    ~ "2. 155-159.9 cm",
    maternal_height_cm >= 150 & maternal_height_cm < 155    ~ "3. 150-154.9 cm",
    maternal_height_cm >= 145 & maternal_height_cm < 150    ~ "4. 145-149.9 cm",
    maternal_height_cm >= 80  & maternal_height_cm < 145    ~ "5. < 145 cm",
    is.na(maternal_height_cm)                               ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "maternal_height_cm")
```


```{r}
# Remove variables starting with "maternal_height_4feet8inches_to_less_5feetinches" if they exist
df <- df %>%
  select(-starts_with("maternal_height_4feet8inches_to_less_5feetinches"))

df <- df %>%
  mutate(maternal_height_4feet8inches_to_less_5feetinches = ifelse(df$maternal_height_cm > 142.24  & df$maternal_height_cm < 152.4, 1, 0) %>% as.factor(),
         maternal_height_less_4feet8inches = ifelse(df$maternal_height_cm < 142.24, 1, 0) %>% as.factor(),
         maternal_height_5_2_feet_taller_inches = ifelse(df$maternal_height_cm > 152.4, 1, 0) %>% as.factor())

recoded_vars <- append(recoded_vars, "maternal_height_4feet8inches_to_less_5feetinches")
```


capture drop maternal_bmi* 
gen maternal_bmi_value = v445/100 
replace maternal_bmi_value = . if maternal_bmi_value > 80 
recode maternal_bmi_value (0/18.49999 = 1 "1. < 18.5") (18.5/24.9999 = 2 "2. 18.5-24.9") ///
(25/60 = 3 "3. 25+"), gen(maternal_bmi)

```{r}
# Remove variables starting with "maternal_bmi" if they exist
df <- df %>%
  select(-starts_with("maternal_bmi"))

# Create a new variable maternal_bmi_value and recode it into maternal_bmi
df <- df %>%
  mutate(maternal_bmi_value = v445 / 100) %>%
  mutate(maternal_bmi_value = ifelse(maternal_bmi_value > 80, NA, maternal_bmi_value)) %>%
  mutate(maternal_bmi = fct_case_when(
    maternal_bmi_value >= 0   & maternal_bmi_value < 18.5  ~ "1. < 18.5",
    maternal_bmi_value >= 18.5 & maternal_bmi_value < 25    ~ "2. 18.5-24.9",
    maternal_bmi_value >= 25   & maternal_bmi_value <= 60   ~ "3. 25+",
    is.na(maternal_bmi_value)                               ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "maternal_bmi")
```

capture drop paternal_height // variable is hb3
gen paternal_height_cm = hb3/10 //height in centimeters
replace paternal_height_cm = . if paternal_height_cm > 999
recode paternal_height_cm (160/230 = 1 "1. 160+ cm") ///
(155/159.999 = 2 "2. 155-159.9 cm") (150/154.999 = 3 "3. 150-154.9 cm") ///
(145/149.999 = 4 "4. 145-149.9 cm") (80/144.999 = 5 "5. < 145 cm"), gen(paternal_height)
qui bysort hhid (paternal_height): replace paternal_height = paternal_height[1]

```{r}
# Remove variables starting with "paternal_height" if they exist
df <- df %>%
  select(-starts_with("paternal_height"))

# Create a new variable paternal_height_cm and recode it into paternal_height
df <- df %>%
  mutate(paternal_height_cm = hb3 / 10) %>%
  mutate(paternal_height_cm = ifelse(paternal_height_cm > 999, NA, paternal_height_cm)) %>%
  mutate(paternal_height = case_when(
    paternal_height_cm >= 160 & paternal_height_cm <= 230   ~ "1. 160+ cm",
    paternal_height_cm >= 155 & paternal_height_cm < 160    ~ "2. 155-159.9 cm",
    paternal_height_cm >= 150 & paternal_height_cm < 155    ~ "3. 150-154.9 cm",
    paternal_height_cm >= 145 & paternal_height_cm < 150    ~ "4. 145-149.9 cm",
    paternal_height_cm >= 80  & paternal_height_cm < 145    ~ "5. < 145 cm",
    is.na(paternal_height_cm)                               ~ NA_character_
  ))

# Replace paternal_height with the first value of paternal_height within each hhid group
df <- df %>%
  group_by(hhid) %>%
  mutate(paternal_height = first(paternal_height)) %>%
  ungroup()

recoded_vars <- append(recoded_vars, "paternal_height")
```

capture drop paternal_bmi // variable is hb40
gen paternal_bmi_value = hb40/100 
replace paternal_bmi_value = . if paternal_bmi_value > 80 
recode paternal_bmi_value (0/18.49999 = 1 "1. < 18.5") (18.5/24.9999 = 2 "2. 18.5-24.9") ///
(25/60 = 3 "3. 25+"), gen(paternal_bmi)
qui bysort hhid (paternal_bmi): replace paternal_bmi = paternal_bmi[1]

```{r}
# Remove variables starting with "paternal_bmi" if they exist
df <- df %>%
  select(-starts_with("paternal_bmi"))

# Create a new variable paternal_bmi_value and recode it into paternal_bmi
df <- df %>%
  mutate(paternal_bmi_value = hb40 / 100) %>%
  mutate(paternal_bmi_value = ifelse(paternal_bmi_value > 80, NA, paternal_bmi_value)) %>%
  mutate(paternal_bmi = fct_case_when(
    paternal_bmi_value >= 0   & paternal_bmi_value < 18.5  ~ "1. < 18.5",
    paternal_bmi_value >= 18.5 & paternal_bmi_value < 25    ~ "2. 18.5-24.9",
    paternal_bmi_value >= 25   & paternal_bmi_value <= 60   ~ "3. 25+",
    is.na(paternal_bmi_value)                               ~ NA_character_
  ))

# Remove the temporary variable paternal_bmi_value
df$paternal_bmi_value <- NULL

# Replace paternal_bmi with the first value of paternal_bmi within each hhid group
df <- df %>%
  group_by(hhid) %>%
  mutate(paternal_bmi = first(paternal_bmi)) %>%
  ungroup()

recoded_vars <- append(recoded_vars, "paternal_bmi")
```

capture drop hhadults
recode hv217 (0/1 = 1 "1. 1 or less adults")(2/3 = 2 "2. 2 adults") (4 = 3 "3. 3+ adults")(5=.), gen(hhadults)

capture drop hh_members
recode hv009 (2/4 = 1 "1. 4 or less")(5/8 = 2 "2. 5 to 8")(9/41  = 3 "3. 9 or more"), gen(hh_members)
*use the variable df for the df

```{r}
# Remove the variables hhadults and hh_members if they exist
if ("hhadults" %in% colnames(df)) {
  df$hhadults <- NULL
}
if ("hh_members" %in% colnames(df)) {
  df$hh_members <- NULL
}

# Create a new variable hhadults and hh_members based on recoded values
df <- df %>%
  mutate(hhadults = fct_case_when(
    hv217 >= 0 & hv217 <= 1 ~ "1. 1 or less adults",
    hv217 >= 2 & hv217 <= 3 ~ "2. 2 adults",
    hv217 == 4              ~ "3. 3+ adults",
    hv217 == 5              ~ NA_character_
  )) %>%
  mutate(hh_members = fct_case_when(
    hv009 >= 2 & hv009 <= 4  ~ "1. 4 or less",
    hv009 >= 5 & hv009 <= 8  ~ "2. 5 to 8",
    hv009 >= 9 & hv009 <= 41 ~ "3. 9 or more"
  ))

recoded_vars <- append(recoded_vars, "hhadults")
recoded_vars <- append(recoded_vars, "hh_members")
```

capture drop m_age_marriage
gen m_age_marriage = v511
capture drop marital_age
recode m_age_marriage (0/17.999 = 2 "2. < 18 Years") (18/60 = 1 "1. 18 or More") (. = . ), gen(marital_age)

capture drop firstbirth_18y
recode v212 (0/17 = 0 "0. < 17 years")(18/45 = 1 "1. 18+")(46/. = .), gen(firstbirth_18y)

* dietary diversity based on Ruel and Menon 2002, from Subramanian 2016 PLOS one 
* includes milk, meat, lentils, starchy staples, vit A fruits, other fruits, dairy, oils/fat/butter
* v408-v414q dietary intake of lastchild
* 


```{r}
# Remove the variables m_age_marriage, marital_age, and firstbirth_18y if they exist
if ("m_age_marriage" %in% colnames(df)) {
  df$m_age_marriage <- NULL
}
if ("marital_age" %in% colnames(df)) {
  df$marital_age <- NULL
}
if ("firstbirth_18y" %in% colnames(df)) {
  df$firstbirth_18y <- NULL
}

# Create new variables marital_age and firstbirth_18y based on recoded values
df <- df %>%
  mutate(m_age_marriage = v511) %>%
  mutate(marital_age = fct_case_when(
    m_age_marriage >= 18 & m_age_marriage <= 60 ~ "1. 18 or More",    
    m_age_marriage >= 0 & m_age_marriage < 18  ~ "2. < 18 Years",
    is.na(m_age_marriage)                       ~ NA_character_
  )) %>%
  mutate(firstbirth_18y = fct_case_when(
    v212 >= 0  & v212 < 17  ~ "0. < 17 years",
    v212 >= 18 & v212 <= 45 ~ "1. 18+",
    v212 >= 46              ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "m_age_marriage")
recoded_vars <- append(recoded_vars, "firstbirth_18y")
```

capture drop still_breastfed
gen still_breastfed = .
replace still_breastfed = 1 if m4 == 95
replace still_breastfed = 0 if m4 != 95 & child_age <=60

```{r}
# Remove the variable still_breastfed if it exists
if ("still_breastfed" %in% colnames(df)) {
  df$still_breastfed <- NULL
}

# Create the new variable still_breastfed based on the given conditions
df <- df %>%
  mutate(still_breastfed = if_else(m4 == 95, 1, if_else(m4 != 95 & child_age <= 60, 0, NA_real_)))

df$still_breastfed <- as.factor(df$still_breastfed)

recoded_vars <- append(recoded_vars, "still_breastfed")
```

capture drop icfi_breastmilk
gen icfi_breastmilk_score = . 
replace icfi_breastmilk_score = 0 if child_age >=6 & child_age <=60 
replace icfi_breastmilk_score = 2 if still_breastfed == 1 & child_age >=6 & child_age <=12
replace icfi_breastmilk_score = 1 if still_breastfed == 1 & child_age >12 & child_age <=60

```{r}
# Remove the variable icfi_breastmilk_score if it exists
if ("icfi_breastmilk_score" %in% colnames(df)) {
  df$icfi_breastmilk_score <- NULL
}

# # Create the new variable icfi_breastmilk_score based on the given conditions
# df <- df %>%
#   mutate(icfi_breastmilk_score = case_when(
#     child_age >= 6 & child_age <= 60 ~ 0,
#     still_breastfed == 1 & child_age >= 6 & child_age <= 12 ~ 2,
#     still_breastfed == 1 & child_age > 12 & child_age <= 60 ~ 1,
#     TRUE ~ NA_real_
#   ))

df$icfi_breastmilk_score <- NA
df$icfi_breastmilk_score[df$child_age >= 6 & df$child_age <= 60 & is.na(df$still_breastfed)] <- 0
df$icfi_breastmilk_score[df$still_breastfed == 1 & df$child_age >= 6 & df$child_age <= 12] <- 2
df$icfi_breastmilk_score[df$still_breastfed == 1 & df$child_age > 12 & df$child_age <= 60] <- 1
df$icfi_breastmilk_score <- as.factor(df$icfi_breastmilk_score)

recoded_vars <- append(recoded_vars, "icfi_breastmilk_score")
```

capture drop icfi_bottlefeed 
gen icfi_bottlefeed_score = . 
replace icfi_bottlefeed_score = 1 if  child_age >=6 & child_age <=60 & m38 == 0 
replace icfi_bottlefeed_score = 0 if  child_age >=6 & child_age <=60 & m38 == 1

```{r}
# Remove the variable icfi_bottlefeed_score if it exists
if ("icfi_bottlefeed_score" %in% colnames(df)) {
  df$icfi_bottlefeed_score <- NULL
}

# Create the new variable icfi_bottlefeed_score based on the given conditions
# df <- df %>%
#   mutate(icfi_bottlefeed_score = if_else(child_age >= 6 & child_age <= 60 & m38 == 0, 1,
#                                          if_else(child_age >= 6 & child_age <= 60 & m38 == 1, 0, NA_real_)))

df$icfi_bottlefeed_score <- NA
df$icfi_bottlefeed_score[df$child_age >= 6 & df$child_age <= 60 & df$m38 == 0] <- 1
df$icfi_bottlefeed_score[df$child_age >= 6 & df$child_age <= 60 & df$m38 == 1] <- 0
df$icfi_bottlefeed_score <- df$icfi_bottlefeed_score %>% as.factor()

recoded_vars <- append(recoded_vars, "icfi_bottlefeed_score")
```

capture drop icfi_starchy_staples_score 
gen icfi_starchy_staples_score = . 
replace icfi_starchy_staples_score = 0 if child_age >=6 & child_age  <= 24 
replace icfi_starchy_staples_score = 1 if (v412a == 1 | v412b == 1 | v414e == 1 | v414f == 1 ) & child_age >=6 & child_age <=60 

```{r}
# Remove the variable icfi_starchy_staples_score if it exists
if ("icfi_starchy_staples_score" %in% colnames(df)) {
  df$icfi_starchy_staples_score <- NULL
}

# # Create the new variable icfi_starchy_staples_score based on the given conditions
# df <- df %>%
#   mutate(icfi_starchy_staples_score = if_else(child_age >= 6 & child_age <= 24, 0,
#                                               if_else((v412a == 1 | v412b == 1 | v414e == 1 | v414f == 1) & child_age >= 6 & child_age <= 60, 1, NA_real_)))

df$icfi_starchy_staples_score <- NA
df$icfi_starchy_staples_score[df$child_age >= 6 & df$child_age <= 24] <- 0
df$icfi_starchy_staples_score[(df$v412a == 1 | df$v412b == 1 | df$v414e == 1 | df$v414f == 1) & df$child_age >= 6 & df$child_age <= 60] <- 1
df$icfi_starchy_staples_score <- df$icfi_starchy_staples_score %>% as.factor()

recoded_vars <- append(recoded_vars, "icfi_starchy_staples_score")
```

capture drop icfi_lentils_score
gen icfi_lentils_score = . 
replace icfi_lentils_score = 0 if child_age >=6 & child_age <=60 
replace icfi_lentils_score = 1 if v414o == 1  & child_age >=6 & child_age <=60 

```{r}
# Remove the variable icfi_lentils_score if it exists
if ("icfi_lentils_score" %in% colnames(df)) {
  df$icfi_lentils_score <- NULL
}

# Create the new variable icfi_lentils_score based on the given conditions
# df <- df %>%
#   mutate(icfi_lentils_score = if_else(child_age >= 6 & child_age <= 60, 0,
#                                       if_else(v414o == 1 & child_age >= 6 & child_age <= 60, 1, NA_real_)))

df$icfi_lentils_score <- NA
df$icfi_lentils_score[df$child_age >= 6 & df$child_age <= 60] <- 0
df$icfi_lentils_score[df$v414o == 1 & df$child_age >= 6 & df$child_age <= 60] <- 1
df$icfi_lentils_score <- df$icfi_lentils_score %>% as.factor()

recoded_vars <- append(recoded_vars, "icfi_lentils_score")
```


capture drop icfi_milk_score
gen icfi_milk_score = . 
replace icfi_milk_score = 0 if child_age >=6 & child_age <=60 
replace icfi_milk_score = 1 if (v411 == 1 ) & child_age >=6 & child_age <=60 

```{r}
# Remove the variable icfi_milk_score if it exists
if ("icfi_milk_score" %in% colnames(df)) {
  df$icfi_milk_score <- NULL
}

# # Create the new variable icfi_milk_score based on the given conditions
# df <- df %>%
#   mutate(icfi_milk_score = if_else(child_age >= 6 & child_age <= 60, 0,
#                                    if_else(v411 == 1 & child_age >= 6 & child_age <= 60, 1, NA_real_)))

df$icfi_milk_score <- NA
df$icfi_milk_score[df$child_age >= 6 & df$child_age <= 60] <- 0
df$icfi_milk_score[df$v411 == 1 & df$child_age >= 6 & df$child_age <= 60] <- 1
df$icfi_milk_score <- as.factor(df$icfi_milk_score)

recoded_vars <- append(recoded_vars, "icfi_milk_score")
```

capture drop icfi_dairy_score
gen icfi_dairy_score = . 
replace icfi_dairy_score = 0 if child_age >=6 & child_age <=60 
replace icfi_dairy_score = 1 if ( v414p == 1 |v414q == 1) & child_age >=6 & child_age <=60 

```{r}
# Remove the variable icfi_dairy_score if it exists
if ("icfi_dairy_score" %in% colnames(df)) {
  df$icfi_dairy_score <- NULL
}

# # Create the new variable icfi_dairy_score based on the given conditions
# df <- df %>%
#   mutate(icfi_dairy_score = if_else(child_age >= 6 & child_age <= 60, 0,
#                                     if_else((v414p == 1 | v414q == 1) & child_age >= 6 & child_age <= 60, 1, NA_real_)))

df$icfi_dairy_score <- NA
df$icfi_dairy_score[df$child_age >= 6 & df$child_age <= 60] <- 0
df$icfi_dairy_score[(df$v414p == 1 | df$v414q == 1) & df$child_age >= 6 & df$child_age <= 60] <- 1
df$icfi_dairy_score <- df$icfi_dairy_score %>% as.factor()

recoded_vars <- append(recoded_vars, "icfi_dairy_score")
```


capture drop icfi_meat_score
gen icfi_meat_score = . 
replace icfi_meat_score = 0 if child_age >=6 & child_age <=60 
replace icfi_meat_score = 1 if (v414a == 1 | v414m == 1) & child_age >=6 & child_age <=60 

```{r}
# Remove the variable icfi_meat_score if it exists
if ("icfi_meat_score" %in% colnames(df)) {
  df$icfi_meat_score <- NULL
}

# # Create the new variable icfi_meat_score based on the given conditions
# df <- df %>%
#   mutate(icfi_meat_score = if_else(child_age >= 6 & child_age <= 60, 0,
#                                    if_else((v414a == 1 | v414m == 1) & child_age >= 6 & child_age <= 60, 1, NA_real_)))

df$icfi_meat_score <- NA
df$icfi_meat_score[df$child_age >= 6 & df$child_age <= 60] <- 0
df$icfi_meat_score[(df$v414a == 1 | df$v414m == 1) & df$child_age >= 6 & df$child_age <= 60] <- 1
df$icfi_meat_score <- as.factor(df$icfi_meat_score)

recoded_vars <- append(recoded_vars, "icfi_meat_score")
```

capture drop icfi_eggs_fish_score
gen icfi_eggs_fish_score = . 
replace icfi_eggs_fish_score = 0 if child_age >=6 & child_age <=60 
replace icfi_eggs_fish_score = 1 if (v414g == 1 | v414n == 1 ) & child_age >=6 & child_age <=60 
replace icfi_meat_score = 1 if icfi_eggs_fish_score == 1

```{r}
# Remove the variable icfi_eggs_fish_score if it exists
if ("icfi_eggs_fish_score" %in% colnames(df)) {
  df$icfi_eggs_fish_score <- NULL
}

# # Create the new variable icfi_eggs_fish_score based on the given conditions
# df <- df %>%
#   mutate(icfi_eggs_fish_score = if_else(child_age >= 6 & child_age <= 60, 0,
#                                         if_else((v414g == 1 | v414n == 1) & child_age >= 6 & child_age <= 60, 1, NA_real_))) %>%
#   mutate(icfi_meat_score = if_else(icfi_eggs_fish_score == 1, 1, icfi_meat_score))

df$icfi_eggs_fish_score <- NA
df$icfi_eggs_fish_score[df$child_age >= 6 & df$child_age <= 60] <- 0
df$icfi_eggs_fish_score[(df$v414g == 1 | df$v414n == 1) & df$child_age >= 6 & df$child_age <= 60] <- 1

# Ensure `icfi_meat_score` is initialized in df
df$icfi_meat_score <- ifelse(is.na(df$icfi_meat_score), NA, df$icfi_meat_score)

# Update `icfi_meat_score` based on `icfi_eggs_fish_score`
df$icfi_meat_score[df$icfi_eggs_fish_score == 1] <- 1

df$icfi_eggs_fish_score <- as.factor(df$icfi_eggs_fish_score)

recoded_vars <- append(recoded_vars, "icfi_eggs_fish_score")
```

capture drop icfi_vitA_score
gen icfi_vitA_score = . 
replace icfi_vitA_score = 0 if child_age >=6 & child_age <=60 
replace icfi_vitA_score = 1 if (v414i == 1 | v414j == 1 | v414k == 1 ) & child_age >=6 & child_age <=60 

```{r}
# Remove the variable icfi_vitA_score if it exists
if ("icfi_vitA_score" %in% colnames(df)) {
  df$icfi_vitA_score <- NULL
}

# # Create the new variable icfi_vitA_score based on the given conditions
# df <- df %>%
#   mutate(icfi_vitA_score = if_else(child_age >= 6 & child_age <= 60, 0,
#                                    if_else((v414i == 1 | v414j == 1 | v414k == 1) & child_age >= 6 & child_age <= 60, 1, NA_real_)))

df$icfi_vitA_score <- NA
df$icfi_vitA_score[df$child_age >= 6 & df$child_age <= 60] <- 0
df$icfi_vitA_score[(df$v414i == 1 | df$v414j == 1 | df$v414k == 1) & df$child_age >= 6 & df$child_age <= 60] <- 1
df$icfi_vitA_score <- as.factor(df$icfi_vitA_score)

recoded_vars <- append(recoded_vars, "icfi_vitA_score")
```

capture drop icfi_other_fruits_score
gen icfi_other_fruits_score = . 
replace icfi_other_fruits_score = 0 if child_age >=6 & child_age <=60 
replace icfi_other_fruits_score = 1 if v414l == 1 & child_age >=6 & child_age <=60 

```{r}
# Remove the variable icfi_other_fruits_score if it exists
if ("icfi_other_fruits_score" %in% colnames(df)) {
  df$icfi_other_fruits_score <- NULL
}

# Create the new variable icfi_other_fruits_score based on the given conditions
# df <- df %>%
#   mutate(icfi_other_fruits_score = if_else(child_age >= 6 & child_age <= 60, 0,
#                                            if_else(v414l == 1 & child_age >= 6 & child_age <= 60, 1, NA_real_)))

df$icfi_other_fruits_score <- NA
df$icfi_other_fruits_score[df$child_age >= 6 & df$child_age <= 60] <- 0
df$icfi_other_fruits_score[df$v414l == 1 & df$child_age >= 6 & df$child_age <= 60] <- 1
df$icfi_other_fruits_score <- as.factor(df$icfi_other_fruits_score)

recoded_vars <- append(recoded_vars, "icfi_other_fruits_score")
```


capture drop dietary_diversity*
egen dietary_diversity_score = rowtotal(icfi_milk_score icfi_meat_score icfi_lentils_score icfi_dairy_score  icfi_starchy_staples_score ///
 icfi_vitA_score icfi_other_fruits_score )
 replace dietary_diversity_score = . if icfi_milk_score == . & icfi_meat_score  == . & ///
 icfi_lentils_score  == . &  icfi_dairy_score  == . &  icfi_starchy_staples_score  == . &  ///
 icfi_vitA_score  == . &  icfi_other_fruits_score  == . 
recode dietary_diversity_score (0 = 1 "quintile 1") (1 = 2 "quintile 2") (2 = 3 "quintile 3") ///
(3/4 = 4 "quintile 4") (5/8 = 5 "quintile 5"), gen (dietary_diversity)

```{r}
# Remove the variable dietary_diversity_score and dietary_diversity if they exist
if ("dietary_diversity_score" %in% colnames(df)) {
  df$dietary_diversity_score <- NULL
}
if ("dietary_diversity" %in% colnames(df)) {
  df$dietary_diversity <- NULL
}

# Calculate the dietary_diversity_score and create the dietary_diversity variable
df <- df %>%
  mutate(dietary_diversity_score = rowSums(cbind(icfi_milk_score, icfi_meat_score, icfi_lentils_score, icfi_dairy_score,
                                                 icfi_starchy_staples_score, icfi_vitA_score, icfi_other_fruits_score),
                                           na.rm = TRUE),
         dietary_diversity = case_when(
           dietary_diversity_score == 0 ~ 1,
           dietary_diversity_score == 1 ~ 2,
           dietary_diversity_score == 2 ~ 3,
           dietary_diversity_score >= 3 & dietary_diversity_score <= 4 ~ 4,
           dietary_diversity_score >= 5 & dietary_diversity_score <= 8 ~ 5,
           is.na(dietary_diversity_score) ~ NA_real_
         ))

# Set dietary_diversity_score to NA if all the other scores are NA
df$dietary_diversity_score[is.na(df$icfi_milk_score) & is.na(df$icfi_meat_score) &
                            is.na(df$icfi_lentils_score) & is.na(df$icfi_dairy_score) &
                            is.na(df$icfi_starchy_staples_score) & is.na(df$icfi_vitA_score) &
                            is.na(df$icfi_other_fruits_score)] <- NA

recoded_vars <- append(recoded_vars, "dietary_diversity_score")
```

xtile dd_quintile = dietary_diversity_score if child_age > = 6 & elig == 1, nq(4)

```{r}
# Assuming dietary_diversity_score, child_age, and elig are vectors in a data frame named 'df'

df <- df %>%
  mutate(dd_quintile = ifelse(child_age >= 6 & elig == 1,
                              {
                                # Calculate unique quantile breaks
                                quantiles <- unique(quantile(dietary_diversity_score[child_age >= 6 & elig == 1],
                                                             probs = seq(0, 1, by = 0.25), na.rm = TRUE))
                                
                                # Check if there are enough unique breaks to create the bins
                                if (length(quantiles) > 1) {
                                  as.numeric(cut(dietary_diversity_score,
                                                 breaks = quantiles,
                                                 include.lowest = TRUE))
                                } else {
                                  NA
                                }
                              },
                              NA))

```

capture drop bf_start
recode v426 (0/101 = 1 "1. within first hour") (102/299 = 2 "2. after 1hr") ///
(99/. = .), gen(bf_start)

```{r}
if ("bf_start" %in% colnames(df)) {
  df$bf_start <- NULL
}

# Create the bf_start variable
df <- df %>%
  mutate(bf_start = fct_case_when(v426 >= 0 & v426 <= 101 ~ "1. within first hour",
                                  v426 >= 102 & v426 <= 299 ~ "2. after 1hr",
                                  v426 == 99 | is.na(v426) ~ NA_character_))

recoded_vars <- append(recoded_vars, "bf_start")
```

capture drop infectious_disease
gen infectious_disease = . 
replace infectious_disease = 0 if h11 == 0 & h31 == 0 
replace infectious_disease = 1 if h11 == 2 | h31 == 2

```{r}
if ("infectious_disease" %in% colnames(df)) {
  df$infectious_disease <- NULL
}

df <- df %>%
  mutate(
    infectious_disease = fct_case_when(
      h11 == 0 & h31 == 0 ~ 0,
      h11 == 2 | h31 == 2 ~ 1,
      TRUE ~ NA_real_
    )
  )

recoded_vars <- append(recoded_vars, "infectious_disease")
```

capture drop water_source*
recode hv201 (11 12  = 1 "1. piped") ///
(13 21 41 71 31 32 42 43 51 61 62 72 = 0 "0. not piped") (96 = . ), gen (water_source)


```{r}
# Remove 'water_source' variable if it exists
if ("water_source" %in% colnames(df)) {
  df$water_source <- NULL
}
# Recode the variable "hv201" and create a new variable called "water_source"
df <- df %>% 
  mutate(water_source = fct_case_when(
    hv201 %in% c(13, 21, 41, 71, 31, 32, 42, 43, 51, 61, 62, 72) ~ "0. not piped",
    hv201 %in% c(11, 12) ~ "1. piped",
    hv201 == 96 ~ NA_character_,
    TRUE ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "water_source")
```

capture drop water_treatment
recode hv237 (0 = 0 "0.no")(1 = 1 "1.yes")(8 = . ), gen(water_treatment)


```{r}
# Remove 'water_treatment' variable if it exists
if ("water_treatment" %in% colnames(df)) {
  df$water_treatment <- NULL
}

# Recode the variable "hv237" and create a new variable called "water_treatment"
df <- df %>% 
  mutate(water_treatment = fct_case_when(
    hv237 == 0 ~ "0.no",
    hv237 == 1 ~ "1.yes",
    hv237 == 8 ~ NA_character_,
    TRUE ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "water_treatment")
```

capture drop stool_disposal
recode v465 (1/5 = 1 "1. proper disposal") (9/99 = 0 "0. improper disposal"), gen(stool_disposal)

```{r}
# Remove 'stool_disposal' variable if it exists
if ("stool_disposal" %in% colnames(df)) {
  df$stool_disposal <- NULL
}

# Recode the variable "v465" and create a new variable called "stool_disposal"
df <- df %>% 
  mutate(stool_disposal = fct_case_when(
    between(v465, 9, 99) ~ "0. improper disposal",
    between(v465, 1, 5) ~ "1. proper disposal",
    TRUE ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "stool_disposal")
```


capture drop toilet_type
recode hv205 (11/15 = 1 "1. flush toilet") (21/23 = 2 "2. pit or dry toilet") ///
(41/96 = 2 "2. pit or dry toilet") (31 = 0 "0. no toilet"), gen(toilet_type)
replace toilet_type = . if child_age > 60 

```{r}
# Remove 'toilet_type' variable if it exists
if ("toilet_type" %in% colnames(df)) {
  df$toilet_type <- NULL
}

# Recode the variable "hv205" and create a new variable called "toilet_type"
df <- df %>% 
  mutate(toilet_type = case_when(
    hv205 == 31 ~ "0. no toilet",    
    between(hv205, 11, 15) ~ "1. flush toilet",
    between(hv205, 21, 23) | between(hv205, 41, 96) ~ "2. pit or dry toilet",
    TRUE ~ NA_character_
  ))

# Replace "toilet_type" with missing values if "child_age" is greater than 60
df$toilet_type[df$child_age > 60] <- NA_character_

df$toliet_access_no <- ifelse(df$toilet_type == "0. no toilet", 1,0)  %>% as.factor()


recoded_vars <- append(recoded_vars, "toilet_type")
recoded_vars <- append(recoded_vars, "toliet_access_no")
```

recode hv225 (0 = 0 "0. no") (1 = 1 "1. yes"), gen(toiletshare)

```{r}
# Recode the variable "hv225" and create a new variable called "toiletshare"
df <- df %>% 
  mutate(toiletshare = fct_case_when(
    hv225 == 0 ~ "0. no",
    hv225 == 1 ~ "1. yes",
    TRUE ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "toiletshare")
```

capture drop fuel_type 
recode hv226 (1/5 = 1 "1. electricity/gas/kerosene") (6/8 = 2 "2. wood/coal") ///
(9/11 = 3 "3. straw/hay/dung") (96 = .), gen (fuel_type)

```{r}
# Remove 'toilet_type' variable if it exists
if ("fuel_type" %in% colnames(df)) {
  df$fuel_type <- NULL
}

# Recode the variable "hv226" and create a new variable called "fuel_type"
df <- df %>% 
  mutate(fuel_type = fct_case_when(
    between(hv226, 1, 5) ~ "1. electricity/gas/kerosene",
    between(hv226, 6, 8) ~ "2. wood/coal",
    between(hv226, 9, 11) ~ "3. straw/hay/dung",
    hv226 == 96 ~ NA_character_,
    TRUE ~ NA_character_
  ))

recoded_vars <- append(recoded_vars, "fuel_type")
```


capture drop cooking_fuel
gen cooking_fuel = . 
replace cooking_fuel = 1 if fuel_type == 1
replace cooking_fuel = 0 if fuel_type == 2 | fuel_type == 3
label define cooking_fuel 0 "0. solid fuel" 1 "1. non-solid fuel"
label value cooking_fuel cooking_fuel
```{r}
# remove cooking_fuel if it exists
if ("cooking_fuel" %in% names(df)) {
  df <- df %>% select(-cooking_fuel)
}

df$cooking_fuel <- case_when(
  df$fuel_type == "1. electricity/gas/kerosene" ~ 1,
  df$fuel_type == "2. wood/coal" | df$fuel_type == "3. straw/hay/dung" ~ 0,
  TRUE ~ NA_real_
) %>% as.factor()

df$cooking_fuel_solid <- ifelse(df$cooking_fuel == 0, 1, 0) %>% as.factor()
recoded_vars <- append(recoded_vars, "cooking_fuel_solid")

levels(df$cooking_fuel) <- c("0. solid fuel", "1. non-solid fuel")
recoded_vars <- append(recoded_vars, "cooking_fuel")
```

capture drop kitchen
gen kitchen = hv242
replace kitchen = 1 if hv241 == 2 | hv241 == 3

```{r}
# Remove 'kitchen' variable if it exists
if ("kitchen" %in% colnames(df)) {
  df$kitchen <- NULL
}

# Create 'kitchen' variable based on 'hv242' and 'hv241'
df <- df %>%
  mutate(kitchen = hv242) %>%
  mutate(kitchen = ifelse(hv241 == 2 | hv241 == 3, 1, kitchen))

df$kitchen <- as.factor(df$kitchen)

recoded_vars <- append(recoded_vars, "kitchen")

df$separate_kitchen_no <- ifelse(df$kitchen == 0, 1,0)  %>% as.factor()
recoded_vars <- append(recoded_vars, "separate_kitchen_no")
```


capture drop air_quality //Basani et al 2010, from Subramanian
gen air_quality = . 
replace air_quality = 1 if fuel_type == 1
replace air_quality = 2 if hv242 == 1 & (fuel_type == 2 |fuel_type == 3)
replace air_quality = 3 if hv242 != 1 & (fuel_type == 2 |fuel_type == 3)
replace air_quality = 2 if (hv241 == 2 | hv241==3) & (fuel_type == 2 | fuel_type == 3)
label define air_quality 1 "1. non-solid fuels" 2 "solid fuels in separate kitchen" ///
3 "solid fuels in non-separate kitchen", replace
replace air_quality = . if child_age > 60 
label value air_quality air_quality

```{r}
# Remove 'air_quality' variable if it exists
if ("air_quality" %in% colnames(df)) {
  df$air_quality <- NULL
}

# Create 'air_quality' variable based on the conditions provided
df <- df %>%
  mutate(air_quality = NA) %>%
  mutate(air_quality = ifelse(fuel_type == "1. electricity/gas/kerosene", 1, air_quality)) %>%
  mutate(air_quality = ifelse(hv242 == 1 & (fuel_type == "2. wood/coal" | fuel_type == "3. straw/hay/dung"), 2, air_quality)) %>%
  mutate(air_quality = ifelse(hv242 != 1 & (fuel_type == "2. wood/coal" | fuel_type == "3. straw/hay/dung"), 3, air_quality)) %>%
  mutate(air_quality = ifelse((hv241 == 2 | hv241 == 3) & (fuel_type == "2. wood/coal" | fuel_type == "3. straw/hay/dung"), 2, air_quality)) %>%
  mutate(air_quality = ifelse(child_age > 60, NA, air_quality))   %>%
  mutate(air_quality = fct_recode(as.factor(air_quality),
                                  "1. non-solid fuels" = "1",
                                  "solid fuels in separate kitchen" = "2",
                                  "solid fuels in non-separate kitchen" = "3"))

recoded_vars <- append(recoded_vars, "air_quality")
```

capture drop vitamin_a
gen vitamin_a = . 
replace vitamin_a = 0 if h34 == 0 
replace vitamin_a = 1 if h34 == 1

```{r}
# Remove 'vitamin_a' variable if it exists
if ("vitamin_a" %in% colnames(df)) {
  df$vitamin_a <- NULL
}

# Create 'vitamin_a' variable based on 'h34'
df <- df %>%
  mutate(vitamin_a = NA) %>%
  mutate(vitamin_a = ifelse(h34 == 0, 0, vitamin_a)) %>%
  mutate(vitamin_a = ifelse(h34 == 1, 1, vitamin_a))

df$vitamin_a <- as.factor(df$vitamin_a)

recoded_vars <- append(recoded_vars, "vitamin_a")
```

ren hw57 child_anemia

```{r}
# Rename 'hw57' variable to 'child_anemia'
df <- df %>%
  rename(child_anemia = hw57) 

df$child_anemia <- as.factor(df$child_anemia)

recoded_vars <- append(recoded_vars, "child_anemia")
```

capture drop full_vaccination
gen full_vaccination = . 
replace full_vaccination = 0 if h2 != . & h3 != . & h4 != . & h5 != . & h6 != . & h7 != . ///
& h8 != . & h9 != . 
replace full_vaccination = 1 if (h2 == 1 | h2 == 2 | h2 == 3) & (h3 == 1 | h3 == 2 | h3 == 3) & ///
(h4 == 1 | h4 == 2 | h4 == 3) & (h5 == 1 | h5 == 2 | h5 == 3) & (h6 == 1 | h6 == 2 | h6 == 3) & ///
(h7 == 1 | h7 == 2 | h7 == 3) & (h8 == 1 | h8 == 2 | h8 == 3) & (h9 == 1 | h9 == 2 | h9 == 3) 

```{r}
# Remove 'full_vaccination' variable if it exists
if ("full_vaccination" %in% colnames(df)) {
  df$full_vaccination <- NULL
}

# Create 'full_vaccination' variable based on the conditions specified
df <- df %>%
  mutate(full_vaccination = NA) %>%
  mutate(full_vaccination = ifelse(!is.na(h2) & !is.na(h3) & !is.na(h4) & !is.na(h5) &
                                      !is.na(h6) & !is.na(h7) & !is.na(h8) & !is.na(h9), 0, full_vaccination)) %>%
  mutate(full_vaccination = ifelse((h2 %in% 1:3) & (h3 %in% 1:3) & (h4 %in% 1:3) & (h5 %in% 1:3) &
                                      (h6 %in% 1:3) & (h7 %in% 1:3) & (h8 %in% 1:3) & (h9 %in% 1:3), 1, full_vaccination))

df$full_vaccination <- as.factor(df$full_vaccination)

recoded_vars <- append(recoded_vars, "full_vaccination")
```

capture drop female
recode hv104 (1 = 0 "0. male") (2 = 1 "1.female"), gen(female)

```{r}
# Remove 'female' variable if it exists
if ("female" %in% colnames(df)) {
  df$female <- NULL
}

# Recode the variable "hv104" and create a new variable called "female"
df <- df %>% 
  mutate(female = case_when(
    hv104 == 1 ~ 0,
    hv104 == 2 ~ 1,
    TRUE ~ NA_real_
  )) %>%
  mutate(female = fct_recode(as.factor(female),
                             "0. male" = "0",
                             "1. female" = "1"))

recoded_vars <- append(recoded_vars, "female")
```

```{r}
# Remove 'sex_male' variable if it exists
if ("sex_male" %in% colnames(df)) {
  df$sex_male <- NULL
}

df <- df %>% 
  mutate(sex_male = ifelse(df$female == "0. male",
                                   1,
                                   0)  %>% as.factor())
df$sex_male

recoded_vars <- append(recoded_vars, "sex_male")
```

capture drop total_children_born
gen total_children_born = v201

```{r}
# Remove 'total_children_born' variable if it exists
if ("total_children_born" %in% colnames(df)) {
  df$total_children_born <- NULL
}

# Create 'total_children_born' variable based on 'v201'
df <- df %>%
  mutate(total_children_born = v201)

recoded_vars <- append(recoded_vars, "total_children_born")
```

capture drop total_living_children
gen total_living_children = v218
recode total_living_children (1 = 0 "0. none")(2 = 1 "1. one")(3 = 2 "2. two") (4 = 3 "3. 3") (5/13 = 4 "4. 4 or more"), gen(siblings)
recode siblings (0/1 = 1 "1. 0 or 1")(2 = 2 "2. two")(3/4 = 3 "3. 3 or more"),gen(siblings_2)
recode siblings (0/1 = 1 "1. 0 or 1")(2/4 = 2 "2. 2 or more"),gen(siblings_bin)
capture drop birth_order
gen birth_order =  hc64


```{r}
# Remove 'total_living_children', 'siblings', 'siblings_2', 'siblings_bin', and 'birth_order' variables if they exist
vars_to_remove <- c("total_living_children", "siblings", "siblings_2", "siblings_bin", "birth_order")
df <- df %>% select(-any_of(vars_to_remove))

# Create the new variables based on the given conditions and recodings
df <- df %>%
  mutate(total_living_children = v218,
         siblings = case_when(
           total_living_children == 1 ~ "0. none",
           total_living_children == 2 ~ "1. one",
           total_living_children == 3 ~ "2. two",
           total_living_children == 4 ~ "3. 3",
           between(total_living_children, 5, 13) ~ "4. 4 or more"
         ),
         siblings = factor(siblings, levels = c("0. none", "1. one", "2. two", "3. 3", "4. 4 or more")),
         siblings_2 = case_when(
           siblings %in% c("0. none", "1. one") ~ "1. 0 or 1",
           siblings == "2. two" ~ "2. two",
           siblings %in% c("3. 3", "4. 4 or more") ~ "3. 3 or more"
         ),
         siblings_2 = factor(siblings_2, levels = c("1. 0 or 1", "2. two", "3. 3 or more")),
         siblings_bin = case_when(
           siblings %in% c("0. none", "1. one") ~ "1. 0 or 1",
           siblings %in% c("2. two", "4. 4 or more") ~ "2. 2 or more"
         ),
         siblings_bin = factor(siblings_bin, levels = c("1. 0 or 1", "2. 2 or more")),
         birth_order = hc64)

recoded_vars <- append(recoded_vars, "total_living_children")
recoded_vars <- append(recoded_vars, "siblings")
recoded_vars <- append(recoded_vars, "siblings_2")
recoded_vars <- append(recoded_vars, "siblings_bin")
recoded_vars <- append(recoded_vars, "birth_order")
```

```{r}
# Remove 'num_siblings_two_or_more' variable if it exists
if ("num_siblings_two_or_more" %in% colnames(df)) {
  df$num_siblings_two_or_more  <- NULL
}

df = df %>% 
  mutate(num_siblings_two_or_more = ifelse(df$siblings_bin == "2. 2 or more", 1, 0)  %>% 
           as.factor())


recoded_vars <- append(recoded_vars, "num_siblings_two_or_more")
```



capture drop child_alive 
gen child_alive =  b5

```{r}
# Remove 'child_alive' variable if it exists
if ("child_alive" %in% colnames(df)) {
  df$child_alive <- NULL
}

# Create 'child_alive' variable based on 'b5'
df <- df %>%
  mutate(child_alive = b5)

df$child_alive <- as.factor(df$child_alive)

recoded_vars <- append(recoded_vars, "child_alive")
```


capture drop last_child_id
gen last_child_id = real_id if midx == 1 
sort mother_id midx  

```{r}
# Remove 'last_child_id' variable if it exists
if ("last_child_id" %in% colnames(df)) {
  df$last_child_id <- NULL
}

# Drop the variable "last_child_id" (if it exists) and create a new variable called "last_child_id"
df <- df %>% 
  mutate(last_child_id = ifelse(midx == 1, real_id, NA_real_)) 

# Sort the df by "mother_id" and "midx"
df <- df %>% 
  arrange(mother_id, midx)

recoded_vars <- append(recoded_vars, "last_child_id")
```


```{r}
# First, drop the variable if it already exists
if ("sibling_under5" %in% names(df)) {
  df$sibling_under5 <- NULL
}

# Now, recode v208 into sibling_under5
df$sibling_under5 <- ifelse(df$v208 == 1, 0,
                                   ifelse(df$v208 == 2, 1,
                                          ifelse(df$v208 == 3, 2,
                                                 ifelse(df$v208 >= 4 & df$v208 <= 6, 3, NA))))
# Label the categories
levels(df$sibling_under5) <- c("0. none", "1. one", "2. two", "3. 3 or more")

recoded_vars <- append(recoded_vars, "sibling_under5")
```


```{r}
# Remove 'pre_birth_interval' variable if it exists
vars_to_remove <- c("pre_birth_interval")
df <- df[, !(colnames(df) %in% vars_to_remove)]

# Recode 'b11' values and create 'pre_birth_interval' column
df <- df %>%
  mutate(pre_birth_interval = case_when(
    (b11 >= 0 & b11 <= 23) ~ "2. < 24 months",
    (b11 >= 24 & b11 <= 400) ~ "1. 24+ or first child"
  ))

# Replace 'pre_birth_interval' with "1. 24+ or first child" if 'siblings' is 0
df <- df %>%
  mutate(pre_birth_interval = ifelse(siblings == 0, "1. 24+ or first child", pre_birth_interval))

# Recode 'b12' values and create 'succeeding_birth_interval' column
df <- df %>%
  mutate(succeeding_birth_interval = case_when(
    (b12 >= 0 & b12 <= 23) ~ "1. < 24 months",
    (b12 >= 24 & b12 <= 400) ~ "0. 24+"
  ))

recoded_vars <- append(recoded_vars, "pre_birth_interval")
```

```{r}
# Remove 'pre_birth_interval_less_than_equal_to_24_months' variable if it exists
if ("pre_birth_interval_less_than_equal_to_24_months" %in% colnames(df)) {
  df$pre_birth_interval_less_than_equal_to_24_months <- NULL
}

df <- df %>%
  mutate(pre_birth_interval_less_than_equal_to_24_months = ifelse(df$pre_birth_interval == "2. < 24 months", 1, 0)  %>% as.factor())

recoded_vars <- append(recoded_vars, "pre_birth_interval_less_than_equal_to_24_months")
```


capture drop  brothers brothers_count
gen brothers_count = v202
replace brothers_count = brothers_count-1 if female == 0 & brothers_count !=0
capture drop brothers_home
recode brothers_count (2 = 2 "2. 2 ")(3/10 = 3 "3. 3 or more"), gen(brothers_home)

capture drop sisters sisters_count
gen sisters_count = v203
replace sisters_count = sisters_count-1 if female == 1 & sisters_count !=0
capture drop sisters_home
recode sisters_count (2 = 2 "2. 2")(3/10 = 3 "3. 3 or more"), gen(sisters_home)

capture drop bpl_card
recode sh58 (0 = 0 "0. no") (1 = 1 "1. yes") (8/. = .), gen(bpl_card)

capture drop preg_deworm
recode m60 (0 = 0 "0.no")(1 = 1 "1.yes")(7/9 = . ), gen(preg_deworm)

capture drop preg_ironyn
recode m45 (0 = 0 "0.no")(1 = 1 "1.yes")(7/9 = . ), gen(preg_ironyn)  

capture drop preg_iron
recode m46 (1/30 = 1 "1. 1 mo or less") (31/90 = 2 "2. 1-3 mo")(91/180 = 3 "3. 3-6 mo")(180/300 = 4 "4. 6+ mo")(997/999 = .), gen(preg_iron)
replace preg_iron = 0 if preg_ironyn == 0

capture drop delivery_place
recode m15 (11/13 = 1 "1. at home") (21/33 = 2 "2. at HC facility") (96 = .), gen(delivery_place)

capture drop child_deworm
recode h43 (0 = 0 "0.no")(1 = 1 "1.yes")(7/9 = . ), gen(child_deworm)  

capture drop child_cough
recode h31 (0 = 0 "0.no")(2 = 1 "1.yes")(8 = . ), gen(child_cough) 

```{r}
df$brothers_count <- df$v202

# Subtract 1 from 'brothers_count' if 'female' is 0 and 'brothers_count' is not 0
df$brothers_count <- ifelse(df$female == 0 & df$brothers_count != 0, df$brothers_count - 1, df$brothers_count)

# Drop 'brothers_home' column and create 'brothers_home' with the recoded values of 'brothers_count'
df$brothers_home <- case_when(
  df$brothers_count == 2 ~ "2. 2",
  df$brothers_count >= 3 & df$brothers_count <= 10 ~ "3. 3 or more",
  TRUE ~ NA_character_
)

recoded_vars <- append(recoded_vars, "brothers_count")
recoded_vars <- append(recoded_vars, "brothers_home")
```

```{r}
# Drop 'sisters' and 'sisters_count' columns and create 'sisters_count' with 'v203' values
df$sisters_count <- df$v203

# Subtract 1 from 'sisters_count' if 'female' is 1 and 'sisters_count' is not 0
df$sisters_count <- ifelse(df$female == 1 & df$sisters_count != 0, df$sisters_count - 1, df$sisters_count)

# Drop 'sisters_home' column and create 'sisters_home' with the recoded values of 'sisters_count'
df$sisters_home <- case_when(
  df$sisters_count == 2 ~ "2. 2",
  df$sisters_count >= 3 & df$sisters_count <= 10 ~ "3. 3 or more",
  TRUE ~ NA_character_
)

recoded_vars <- append(recoded_vars, "sisters_count")
recoded_vars <- append(recoded_vars, "sisters_home")
```


```{r}
# sh58 variable (Household has a BPL card) is present in the 2015-16 survey but not in the 2019-21
# Drop 'bpl_card' column and create 'bpl_card' with the recoded values of 'sh58'
#df$bpl_card <- case_when(
#  df$sh58 == 0 ~ "0. no",
#  df$sh58 == 1 ~ "1. yes",
#  is.na(df$sh58) ~ NA_character_,
#  TRUE ~ NA_character_
#)
```


```{r}
# Drop 'preg_deworm' column and create 'preg_deworm' with the recoded values of 'm60'
df$preg_deworm <- case_when(
  df$m60 == 0 ~ "0.no",
  df$m60 == 1 ~ "1.yes",
  df$m60 >= 7 & df$m60 <= 9 ~ NA_character_,
  TRUE ~ NA_character_
)

# Recode 'm45' variable and create 'preg_ironyn' variable
df <- df %>%
  mutate(preg_ironyn = case_when(
    m45 == 0 ~ "0.no",
    m45 == 1 ~ "1.yes",
    m45 >= 7 & m45 <= 9 ~ NA_character_,
    TRUE ~ NA_character_
  ))

# Recode 'm46' variable and create 'preg_iron' variable
df <- df %>%
  mutate(preg_iron = case_when(
    m46 >= 1 & m46 <= 30 ~ "1. 1 mo or less",
    m46 >= 31 & m46 <= 90 ~ "2. 1-3 mo",
    m46 >= 91 & m46 <= 180 ~ "3. 3-6 mo",
    m46 >= 180 & m46 <= 300 ~ "4. 6+ mo",
    m46 >= 997 & m46 <= 999 ~ NA_character_,
    TRUE ~ NA_character_
  ))

# Replace values in 'preg_iron' with 0 where 'preg_ironyn' is 0
df$preg_iron <- ifelse(df$preg_ironyn == "0.no", 0, df$preg_iron)


# Recode 'm15' variable and create 'delivery_place' variable
df <- df %>%
  mutate(delivery_place = case_when(
    m15 >= 11 & m15 <= 13 ~ "1. at home",
    m15 >= 21 & m15 <= 33 ~ "2. at HC facility",
    m15 == 96 ~ NA_character_,
    TRUE ~ NA_character_
  ))

# Recode 'h43' variable and create 'child_deworm' variable
df <- df %>%
  mutate(child_deworm = case_when(
    h43 == 0 ~ "0.no",
    h43 == 1 ~ "1.yes",
    h43 >= 7 & h43 <= 9 ~ NA_character_,
    TRUE ~ NA_character_
  ))

df <- df  %>% 
  mutate(child_cough = case_when(h31 == 0 ~ 0, 
                                 h31 == 2 ~ 1, 
                                 TRUE ~ NA_real_))

df$child_cough <- as.factor(df$child_cough)

recoded_vars <- append(recoded_vars, "preg_deworm")
recoded_vars <- append(recoded_vars, "preg_ironyn")
recoded_vars <- append(recoded_vars, "preg_iron")
recoded_vars <- append(recoded_vars, "delivery_place")
recoded_vars <- append(recoded_vars, "child_deworm")
recoded_vars <- append(recoded_vars, "child_cough")
```


capture drop child_fever
recode h22 (0 = 0 "0.no")(1= 1 "1.yes")(8 = . ), gen(child_fever) 

```{r}
# Drop variables starting with 'child_fever'
if ("child_fever" %in% names(df)) {
  df <- df %>% select(-starts_with("child_fever"))
}

# Recode 'h22' variable into 'child_fever' variable
df <- df %>% mutate(child_fever = case_when(
  h22 == 0 ~ "0.no",
  h22 == 1 ~ "1.yes",
  h22 == 8 ~ NA_character_, # missing
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "child_fever")
```


capture drop child_diarrhea
recode h11 (0 = 0 "0.no")(2 = 1 "1.yes")(8 = . ), gen(child_diarrhea)  

```{r}
# Drop variables starting with 'child_diarrhea'
if ("child_diarrhea" %in% names(df)) {
  df <- df %>% select(-starts_with("child_diarrhea"))
}

# Recode 'h11' variable into 'child_diarrhea' variable
df <- df %>% mutate(child_diarrhea = case_when(
  h11 == 0 ~ "0.no",
  h11 == 2 ~ "1.yes",
  h11 == 8 ~ NA_character_, # missing
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "child_diarrhea")
```


capture drop birthweight
recode m19 (500/1800 = 1 "1. 1800g or less")(1801/2500 = 2 "2. 1801 to 2500g")(2501/9000 = 3 "3. more than 2500g")(9990/. = .),gen(birthweight)

```{r}
# Drop variables starting with 'birthweight'
if ("birthweight" %in% names(df)) {
  df <- df %>% select(-starts_with("birthweight"))
}

# Recode 'm19' variable into 'birthweight' variable
df <- df %>% mutate(birthweight = case_when(
  m19 >= 500 & m19 <= 1800 ~ "1. 1800g or less",
  m19 >= 1801 & m19 <= 2500 ~ "2. 1801 to 2500g",
  m19 >= 2501 & m19 <= 9000 ~ "3. more than 2500g",
  m19 == 9990 ~ NA_character_, # missing
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "birthweight")
```

```{r}
# Drop variables starting with 'birth_weight_less_than_1800_g', `birth_weight_1800_g_2500_g`
if ("birth_weight_less_than_1800_g" %in% names(df)) {
  df <- df %>% select(-starts_with("birth_weight_less_than_1800_g"))
}

if ("birth_weight_1800_g_2500_g" %in% names(df)) {
  df <- df %>% select(-starts_with("birth_weight_1800_g_2500_g"))
}

df <- df %>%
  mutate(birth_weight_less_than_1800_g = ifelse(df$birthweight == "1. 1800g or less", 1, 0)  %>% as.factor(),
         birth_weight_1800_g_2500_g = ifelse(df$birthweight == "2. 1801 to 2500g", 1, 0) %>% as.factor(),
         birth_weight_more_than_2500_g = ifelse(df$birthweight == "3. more than 2500g", 1, 0) %>% as.factor())

recoded_vars <- append(recoded_vars, "birth_weight_less_than_1800_g")
recoded_vars <- append(recoded_vars, "birth_weight_1800_g_2500_g")
recoded_vars <- append(recoded_vars, "birth_weight_more_than_2500_g")
```


capture drop multiplet
recode b0 (10 = 0 "0. no") (1/3 = 1 "1. yes"), gen(multiplet)

```{r}
# Drop variables starting with 'multiplet'
if ("multiplet" %in% names(df)) {
  df <- df %>% select(-starts_with("multiplet"))
}

# Recode 'b0' variable into 'multiplet' variable
df <- df %>% mutate(multiplet = case_when(
  b0 == 10 ~ "0. no",
  b0 %in% 1:3 ~ "1. yes",
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "multiplet")
```


capture drop caste
recode sh36 (1 = 4 "4. Scheduled Caste")(2 = 3 "3. Scheduled Tribe")(3 = 2 "Other backward caste") (4 = 1 "None")(8=.), gen(caste)
replace caste = 1 if sh35 == 993 | sh35 == 998

```{r}
# Variables sh35/sh36 in 2015-16 survey (caste or tribe of household head, Type of caste or tribe of the household head) were replaced by SH48 and SH49 in 2019-21 

# Drop variables starting with 'caste'
if ("caste" %in% names(df)) {
  df <- df %>% select(-starts_with("caste"))
}

# Recode 'sh49' variable into 'caste' variable
df <- df %>% mutate(caste = fct_case_when(
  sh49 == 1 ~ "4. Scheduled Caste",
  sh49 == 2 ~ "3. Scheduled Tribe",
  sh49 == 3 ~ "2. Other backward caste",
  sh49 == 4 ~ "1. None",
  sh49 == 8 ~ NA_character_, # missing
  TRUE ~ NA_character_ # all other cases
))

# Replace 'caste' value with 1 if 'sh48' equals 993 or 998
df <- df %>% mutate(caste = if_else(sh48 %in% c(993, 998), "1. None", caste))

recoded_vars <- append(recoded_vars, "caste")
```


capture drop religion
recode v130 (1 4/8 = 1 "1. Hindu+") (2 = 2 "2. muslim")(3 = 3 "3. christian") (9/96 = 4 "4. other"), gen(religion)

```{r}
# Drop variables starting with 'religion'
if ("religion" %in% names(df)) {
  df <- df %>% select(-starts_with("religion"))
}

# Recode 'v130' variable into 'religion' variable
df <- df %>% mutate(religion = case_when(
  v130 %in% c(1, 4:8) ~ "1. Hindu+",
  v130 == 2 ~ "2. Muslim",
  v130 == 3 ~ "3. Christian",
  v130 >= 9 & v130 <= 96 ~ "4. Other",
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "religion")
```


capture drop rural
recode v025 (1 = 0 "0. urban")(2 = 1 "1. rural"), gen(rural)

```{r}
# Drop variables starting with 'rural'
if ("rural" %in% names(df)) {
  df <- df %>% select(-starts_with("rural"))
}

# Recode 'v025' variable into 'rural' variable
df <- df %>% mutate(rural = case_when(
  v025 == 1 ~ "0. urban",
  v025 == 2 ~ "1. rural",
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "rural")
```


capture drop mother_education 
recode v133 (0 = 0 "0. no education") (1/7 = 1 "1. primary") ///
(8/10 = 2 "2. secondary") (11/12 = 3 "3. higher secondary") (13/22 = 4 "4. college") (99 = .), gen (mother_education)


```{r}
# Drop variables starting with 'mother_education'
if ("mother_education" %in% names(df)) {
  df <- df %>% select(-starts_with("mother_education"))
}

# Create 'mother_education' column based on 'v133' values
df$mother_education <- ifelse(df$v133 == 0, 0,
                              ifelse(df$v133 >= 1 & df$v133 <= 7, 1,
                                     ifelse(df$v133 >= 8 & df$v133 <= 10, 2,
                                            ifelse(df$v133 >= 11 & df$v133 <= 12, 3,
                                                   ifelse(df$v133 >= 13 & df$v133 <= 22, 4, NA
                                                   )))))

recoded_vars <- append(recoded_vars, "mother_education")
```

```{r}
# Drop variables starting with 'maternal_edu_no' 'maternal_edu_1_7' 'maternal_edu_greater_7_10'
if ("maternal_edu_no" %in% names(df)) {
  df <- df %>% select(-starts_with("maternal_edu_no"))
}

if ("maternal_edu_1_7" %in% names(df)) {
  df <- df %>% select(-starts_with("maternal_edu_1_7"))
}

if ("maternal_edu_greater_7_10" %in% names(df)) {
  df <- df %>% select(-starts_with("maternal_edu_greater_7_10"))
}


df <- df %>% 
  mutate(maternal_edu_no = ifelse(df$v133 == 0, 1, 0) %>% as.factor(),
         maternal_edu_1_7 = ifelse(df$v133 >= 1 & df$v133 <= 7,
                              1, 0) %>% as.factor(),
         maternal_edu_greater_7_10 = ifelse(df$v133 >= 7 & df$v133 <= 10,
                              1, 0) %>% as.factor(),
         maternal_edu_greater_10 = ifelse(df$v133 > 10,
                              1, 0) %>% as.factor())

recoded_vars <- append(recoded_vars, "maternal_edu_no")
```


capture drop season //designated by India Meteorological Department http://imd.gov.in/section/nhac/wxfaq.pdf
recode v006 (1/3 12 = 1 "1. Winter")(4/6 = 2 "2. Summer") (7/9 = 3 "3. Monsoon") (10/11 = 4 "4. Post-Monsoon"), gen(season)

```{r}
# Drop variables starting with 'season'
if ("season" %in% names(df)) {
  df <- df %>% select(-starts_with("season"))
}

# Recode 'v006' variable into 'season' variable
df <- df %>% mutate(season = case_when(
  v006 %in% c(1, 2, 3, 12) ~ "1. Winter",
  v006 %in% 4:6 ~ "2. Summer",
  v006 %in% 7:9 ~ "3. Monsoon",
  v006 %in% 10:11 ~ "4. Post-Monsoon",
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "season")
```

ren s564 anganwadi_support

```{r}
## Variable s564 (While breastfeeding, received benefits from ang) in 2015-16 appears to be renamed to S567 (When you were breastfeeding child did you recei) in 2019-21

# Rename variable 's567' to 'anganwadi_support'
df <- df %>% rename(anganwadi_support = s567)

recoded_vars <- append(recoded_vars, "anganwadi_support")
```


capture drop anc_visits
recode m14 (0 = 0 "0. no visits")(1/3 = 1 "1. 1-3 visits")(4/30 = 2 "2. 4 or more visits")(98=.), gen(anc_visits)"

```{r}
# Drop variables starting with 'anc_visits'
if ("anc_visits" %in% names(df)) {
  df <- df %>% select(-starts_with("anc_visits"))
}


# Recode 'm14' variable into 'anc_visits' variable
df <- df %>% mutate(anc_visits = case_when(
  m14 == 0 ~ "0. no visits",
  m14 >= 1 & m14 <= 3 ~ "1. 1-3 visits",
  m14 >= 4 & m14 <= 30 ~ "2. 4 or more visits",
  m14 == 98 ~ NA_character_, # missing
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "anc_visits")
```


/*capture drop covariate_nonmissing
gen covariate_nonmissing = 0
replace covariate_nonmissing = 1 if  wealth_quintile != . & mother_education !=. ///
& maternal_height !=. & maternal_bmi !=. & marital_age!= . & dietary_diversity !=. & bf_start !=. ///
& infectious_disease !=. & water_source !=. & toilet_type !=. & air_quality !=. & iodized_salt !=. /// 
& full_vaccination !=. & vitamin_a !=.

```{r}
# Unknown variable: Ionized Salt

# Create a new variable 'covariate_nonmissing' and initialize it to 0
df$covariate_nonmissing <- 0

# Use 'ifelse()' to set 'covariate_nonmissing' to 1 where all specified variables are non-missing
df$covariate_nonmissing <- ifelse(
  !is.na(df$wealth_quintile) & !is.na(df$mother_education) &
  !is.na(df$maternal_height) & !is.na(df$maternal_bmi) &
  !is.na(df$marital_age) & !is.na(df$dietary_diversity) &
  !is.na(df$bf_start) & !is.na(df$infectious_disease) &
  !is.na(df$water_source) & !is.na(df$toilet_type) &
  !is.na(df$air_quality) & 
  #!is.na(df$iodized_salt) & # Ionized salt variable not found
  !is.na(df$full_vaccination) & !is.na(df$vitamin_a),
  1, df$covariate_nonmissing
)

df$covariate_nonmissing <- df$covariate_nonmissing %>% as.factor()

recoded_vars <- append(recoded_vars, "covariate_nonmissing")
```


* decisionmaking 743a-v743f; v43e not measured
* domestic violence: v744a v744b v744c v744d v744e s936f s936g
* rough sample size under 25 y, last child = 17k
capture drop fem_aut*
recode v743a (1 = 1)(2 = 0.5)(4/6=0), gen(fem_aut_hc)
recode v743b (1 = 1)(2 = 0.5)(4/6=0), gen(fem_aut_large)
recode v743d (1 = 1)(2 = 0.5)(4/6=0), gen(fem_aut_daily)
recode v743f (1 = 1)(2 = 0.5)(4/6=0), gen(fem_aut_visits)
qui bysort hhid (fem_aut_hc): replace fem_aut_hc = fem_aut_hc[1]
qui bysort hhid (fem_aut_large): replace fem_aut_large = fem_aut_large[1]
qui bysort hhid (fem_aut_daily): replace fem_aut_daily = fem_aut_daily[1]
qui bysort hhid (fem_aut_visits): replace fem_aut_visits = fem_aut_visits[1]


```{r}
df <- df %>%
  select(-starts_with("fem_aut")) %>%
  mutate(fem_aut_hc = case_when(v743a == 1 ~ 1, v743a == 2 ~ 0.5, between(v743a, 4, 6) ~ 0),
         fem_aut_large = case_when(v743b == 1 ~ 1, v743b == 2 ~ 0.5, between(v743b, 4, 6) ~ 0),
         fem_aut_daily = case_when(v743d == 1 ~ 1, v743d == 2 ~ 0.5, between(v743d, 4, 6) ~ 0),
         fem_aut_visits = case_when(v743f == 1 ~ 1, v743f == 2 ~ 0.5, between(v743f, 4, 6) ~ 0)) %>%
  group_by(hhid) %>%
  mutate(fem_aut_hc = fem_aut_hc[1],
         fem_aut_large = fem_aut_large[1],
         fem_aut_daily = fem_aut_daily[1],
         fem_aut_visits = fem_aut_visits[1]) %>%
  ungroup() %>%
  select(-starts_with("v743"))

recoded_vars <- append(recoded_vars, "fem_aut_hc")
recoded_vars <- append(recoded_vars, "fem_aut_large")
recoded_vars <- append(recoded_vars, "fem_aut_daily")
recoded_vars <- append(recoded_vars, "fem_aut_visits")
```


capture drop state
gen state = hv024
label value state HV024
capture drop state_ciaf_rank
recode state (15 = 1 "1. Jharkhand") (8 = 2 "2. Dadra and Nagar")(5 = 3 "3. Bihar")(19 = 4 "4. Madhya Pradesh") ///
(33 = 5 " 5. Uttar Pradesh") (11 = 6 "6. Gujarat")(16 = 7 "7. Karnataka")(7 = 8 "8.Chattisgarh")(29 = 9 "9. Rajasthan") ///
(20 = 10 "10. Maharashtra")(22 = 11 "11. Meghalaya")(12 = 12 "12. Haryana")(26 = 13 "13. Odisha")(34 = 14 "14. Uttarakhand") ///
(4 = 15 "15. Assam")(35 = 16 "16. West Bengal")(2 = 17 "17. Andhra Pradesh")(27 = 18 "18. Puducherry")(25 = 19 "19. Delhi") ///
(31 = 20 "20. Tamil Nadu")(9 = 21 "21. Daman and Diu")(3 = 22 "22. Arunachal Pradesh") (30 = 23 "23. Sikkim")(36 = 24 "24. Telangana") ///
(1 = 25 "25. Andaman and Nicobar")(10 = 26 "26. Goa")(28 = 27 "27. Punjab")(13 = 28 "28. Himachal Pradesh")(32 = 29 "29. Tripura") ///
(18 = 30 "30. Lakshwadeep") (14 = 31 "31. Jammu and Kashmir")(6 = 32 "32. Chandigarh")(17 = 33 "33. Kerala")(24 = 34 "34. Nagaland") ///
(21 = 35 "35. Manipur")(23 = 36 "36. Mizoram"), gen(state_ciaf_rank)
capture drop sibling_under5
recode v208 (1 = 0 "0. none")(2 = 1 "1. one")(3 = 2 "2. two") (4/6 = 3 "3. 3 or more"), gen(sibling_under5)

```{r}
# Remove 'state' and 'state_ciaf_rank' variables if they exist
vars_to_remove <- c("state", "state_ciaf_rank")
df <- df[, !(colnames(df) %in% vars_to_remove)]

# Create the new 'state' variable based on the given conditions and recodings
df <- df %>%
  mutate(state = factor(hv024))

# Recode state values using case_when
df <- df %>%
  mutate(state_ciaf_rank = case_when(
    state == 15 ~ "1. Jharkhand",
    state == 8  ~ "2. Dadra and Nagar",
    state == 5  ~ "3. Bihar",
    state == 19 ~ "4. Madhya Pradesh",
    state == 33 ~ "5. Uttar Pradesh",
    state == 11 ~ "6. Gujarat",
    state == 16 ~ "7. Karnataka",
    state == 7  ~ "8. Chattisgarh",
    state == 29 ~ "9. Rajasthan",
    state == 20 ~ "10. Maharashtra",
    state == 22 ~ "11. Meghalaya",
    state == 12 ~ "12. Haryana",
    state == 26 ~ "13. Odisha",
    state == 34 ~ "14. Uttarakhand",
    state == 4  ~ "15. Assam",
    state == 35 ~ "16. West Bengal",
    state == 2  ~ "17. Andhra Pradesh",
    state == 27 ~ "18. Puducherry",
    state == 25 ~ "19. Delhi",
    state == 31 ~ "20. Tamil Nadu",
    state == 9  ~ "21. Daman and Diu",
    state == 3  ~ "22. Arunachal Pradesh",
    state == 30 ~ "23. Sikkim",
    state == 36 ~ "24. Telangana",
    state == 1  ~ "25. Andaman and Nicobar",
    state == 10 ~ "26. Goa",
    state == 28 ~ "27. Punjab",
    state == 13 ~ "28. Himachal Pradesh",
    state == 32 ~ "29. Tripura",
    state == 18 ~ "30. Lakshwadeep",
    state == 14 ~ "31. Jammu and Kashmir",
    state == 6  ~ "32. Chandigarh",
    state == 17 ~ "33. Kerala",
    state == 24 ~ "34. Nagaland",
    state == 21 ~ "35. Manipur",
    state == 23 ~ "36. Mizoram"
  ))

recoded_vars <- append(recoded_vars, "state")
recoded_vars <- append(recoded_vars, "state_ciaf_rank")
```


capture drop state_zones
gen state_zones = . 
replace state_zones = 2 if v024 == 14 | v024 == 28 | v024 == 13 | v024 == 6 | v024 == 29 | v024 == 25 | v024 == 12
replace state_zones = 1 if v024 == 4 | v024 == 3 | v024 == 21 | v024 == 22 | v024 == 23 | v024 == 24 | v024 == 32 | v024 == 30
replace state_zones = 3 if v024 == 7 | v024 == 19 | v024 == 34 | v024 ==  33
replace state_zones = 4 if v024 == 5 | v024 == 15 | v024 == 26 | v024 == 35
replace state_zones = 5 if v024 == 8 | v024 == 9 | v024 == 10 | v024 == 11 | v024 == 20
replace state_zones = 6 if v024 == 2 | v024 == 16 | v024 == 17 | v024 == 18 | v024 == 27 | v024 == 31 | v024 == 36 | v024 == 1
label define state_zones 1 "1. North-Eastern" 2 "2. North" 3 "3. Central" 4 "4. Eastern" 5 "5.Western" 6 "6. Southern"
label value state_zones state_zones


```{r}
# Remove the "state_zones" column if it exists
if ("state_zones" %in% names(df)) {
  df$state_zones <- NULL
}

# Create a new "state_zones" column by recoding the "v024" column
df$state_zones <- case_when(
  df$v024 %in% c(14, 28, 13, 6, 29, 25, 12) ~ 2, # North
  df$v024 %in% c(4, 3, 21, 22, 23, 24, 32, 30) ~ 1, # North-Eastern
  df$v024 %in% c(7, 19, 34, 33) ~ 3, # Central
  df$v024 %in% c(5, 15, 26, 35) ~ 4, # Eastern
  df$v024 %in% c(8, 9, 10, 11, 20) ~ 5, # Western
  df$v024 %in% c(2, 16, 17, 18, 27, 31, 36, 1) ~ 6 # Southern
)

# Label the categories of the "state_zones" column
label_values <- c("1. North-Eastern", "2. North", "3. Central", "4. Eastern", "5. Western", "6. Southern")
names(label_values) <- 1:6
df$state_zones <- factor(df$state_zones, levels = 1:6, labels = label_values)

recoded_vars <- append(recoded_vars, "state_zones")
```

capture drop eag_state
gen eag_state = 0
replace eag_state = 1 if state == 4 | state == 5 | state == 6 | state == 15 | ///
state == 19 | state ==  26 | state == 29 | state == 33 | state == 34 

```{r}
# Remove the "eag_state" column if it exists
if ("eag_state" %in% names(df)) {
  df$eag_state <- NULL
}

# Create a new "eag_state" column by recoding the "state" column
df$eag_state <- ifelse(
  df$state %in% c(4, 5, 6, 15, 19, 26, 29, 33, 34),
  1,
  0
)

recoded_vars <- append(recoded_vars, "eag_state")
```


capture drop priority_state
gen priority_state = 1
replace priority_state = 0 if state == 5 | state == 7 | state == 13 | state == 14 | ///
state == 15 | state == 19 | state == 26 | state == 29 | state == 33 | state == 34
replace priority_state = 2 if state == 3 | state == 4 | state == 21 | state == 22 ///
 | state == 23 | state == 24 | state == 30 | state == 32
 replace priority_state = 3 if state == 1 | state == 6 | state == 8 | state == 9 ///
  | state == 18  | state == 25  | state == 27
recode priority_state (0 2 = 1 "1. High Priority")(1 3 = 0 "0. Normal Priority"), gen(priority_state_bin)
recode priority_state (0 = 2 "2. Other focus")(2 = 1 "1. NE Focus")(1 3 = 0 "0. Normal Focus"), gen(focus_states)

```{r}
# Drop priority_state if it exists
if ("priority_state" %in% names(df)) {
  df$priority_state <- NULL
}

# Generate priority_state variable
df$priority_state <- 1
df$priority_state[df$state %in% c(5, 7, 13, 14, 15, 19, 26, 29, 33, 34)] <- 0
df$priority_state[df$state %in% c(3, 4, 21, 22, 23, 24, 30, 32)] <- 2
df$priority_state[df$state %in% c(1, 6, 8, 9, 18, 25, 27)] <- 3

# Recode priority_state
df$priority_state_bin <- case_when(
  df$priority_state %in% c(0, 2) ~ 1,
  df$priority_state %in% c(1, 3) ~ 0
)

df$focus_states <- case_when(
  df$priority_state == 0 ~ "0. Normal Focus",
  df$priority_state == 1 ~ "1. NE Focus",
  df$priority_state %in% c(2, 3) ~ "2. Other focus"
)

recoded_vars <- append(recoded_vars, "priority_state")
recoded_vars <- append(recoded_vars, "priority_state_bin")
recoded_vars <- append(recoded_vars, "focus_states")
```

```{r}
df$Northeast_focus <- ifelse(df$focus_states == "1. NE Focus", 1, 0)  %>% as.factor()
df$Other_Focus <- ifelse(df$focus_states == "2. Other focus", 1, 0)  %>% as.factor()

recoded_vars <- append(recoded_vars, "Northeast_focus")
recoded_vars <- append(recoded_vars, "Other_Focus")
```


capture drop water_home
gen water_home = .
replace water_home = 1 if hv204 == 0 | hv204 == 996
replace water_home = 0 if hv204 > 0 & hv204 < 990
label define yesno 0 "0.no" 1 "1.yes", replace
label value water_home yesno

```{r}
# remove water_home if it exists
if ("water_home" %in% names(df)) {
  df <- df %>% select(-water_home)
}

# create water_home variable
df <- df %>% 
  mutate(water_home = case_when(
    hv204 == 0 | hv204 == 996 ~ 1,
    hv204 > 0 & hv204 < 990 ~ 0,
    TRUE ~ NA_real_
  ))

# label values
df$water_home <- factor(df$water_home, levels = c(0, 1), labels = c("0.no", "1.yes"))

recoded_vars <- append(recoded_vars, "water_home")
```

capture drop floor_finish
gen floor_finish = . 
replace floor_finish = 1 if hv213 >= 31 
replace floor_finish = 0 if hv213 < 31
label value floor_finish yesno

```{r}
# remove floor_finish if it exists
if ("floor_finish" %in% names(df)) {
  df <- df %>% select(-floor_finish)
}

df$floor_finish <- case_when(
  df$hv213 >= 31 ~ 1,
  df$hv213 < 31 ~ 0,
  TRUE ~ NA_real_
)

levels(df$floor_finish) <- c("0.no", "1.yes")

recoded_vars <- append(recoded_vars, "floor_finish")
```

capture drop wall_finish
gen wall_finish = . 
replace wall_finish = 1 if hv214 >= 31 
replace wall_finish = 0 if hv214 < 31
label value wall_finish yesno


```{r}
# remove floor_finish if it exists
if ("wall_finish" %in% names(df)) {
  df <- df %>% select(-wall_finish)
}

df <- df %>%
  mutate(wall_finish = case_when(
    hv214 >= 31 ~ 1,
    hv214 < 31 ~ 0,
    TRUE ~ NA_real_)) %>%
  mutate(wall_finish = factor(wall_finish, levels = c(0, 1), labels = c("0.no", "1.yes")))

recoded_vars <- append(recoded_vars, "wall_finish")
```


capture drop roof_finish
gen roof_finish = 0 
replace roof_finish = 1 if hv215 == 31 
label value roof_finish yesno

```{r}
# remove roof_finish if it exists
if ("roof_finish" %in% names(df)) {
  df <- df %>% select(-roof_finish)
}

df <- df %>%
  mutate(roof_finish = case_when(
    hv215 == 31 ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(roof_finish = factor(roof_finish, levels = c(0, 1), labels = c("0.no", "1.yes")))

recoded_vars <- append(recoded_vars, "roof_finish")
```

capture drop house_finish
gen house_finish = 0 
replace house_finish = 1 if wall_finish == 1 | roof_finish == 1 | floor_finish == 1
replace house_finish = 2 if wall_finish == 1 & roof_finish == 1 & floor_finish == 1
recode house_finish (2 = 1),gen(house_finish_bin)


```{r}
# remove house_finish if it exists
if ("house_finish" %in% names(df)) {
  df <- df %>% select(-house_finish)
}

df <- df %>%
  mutate(house_finish = case_when(
    wall_finish == 1 | roof_finish == 1 | floor_finish == 1 ~ 1,
    wall_finish == 1 & roof_finish == 1 & floor_finish == 1 ~ 2,
    TRUE ~ 0),
    house_finish_bin = case_when(
      house_finish == 2 ~ 1,
      TRUE ~ house_finish)
  )

# Create label for 'house_finish' variable
df$house_finish <- factor(df$house_finish, levels = c(0, 1, 2),
                          labels = c("0. no concrete material", "1. some concrete material", "2. all concrete"))

recoded_vars <- append(recoded_vars, "house_finish")
```

capture drop use_soap
gen use_soap = . 
replace use_soap = 0 if hv230a < 3
replace use_soap = 1 if  hv232 == 1

```{r}
# # remove use_finish if it exists
if ("use_soap" %in% names(df)) {
  df <- df %>% select(-use_soap)
}

# df$hv232 <- as.numeric(df$hv232)
# 
# df$use_soap <- NA
# df$use_soap <- case_when(
#     df$hv230a < 3 & !is.na(df$hv232)  ~ 0,
#     is.na(df$hv230a) & df$hv232 == 1  ~ 1,
# TRUE ~ NA_real_
# )
# 
# df$use_soap <- case_when(is.na(df$use_soap)  ~ 1,
#                          TRUE ~ 0) %>% as.factor()

df$use_soap <- NA
df$use_soap[df$hv230a < 3] <- 0
df$use_soap[df$hv232 == 1] <- 1
df$use_soap <- df$use_soap %>% as.factor()

recoded_vars <- append(recoded_vars, "use_soap")
```

capture drop reading
gen reading = v155
replace reading = 1 if v155 == 2
replace reading = . if v155 > 2
 
```{r}
# remove reading if it exists
#if ("reading" %in% names(df)) {
#  df <- df %>% select(-reading)
#}

#df <- df %>% 
#  mutate(
#    reading = case_when(
#      v155 == 2 ~ 1,
#      v155 > 2 ~ NA_real_,
#      TRUE ~ v155
#    )
#  )

```

capture drop christian 
gen christian = . 
replace christian = 1 if religion == 3
replace christian = 0 if religion < 3

```{r}
# remove christian if it exists
if ("christian" %in% names(df)) {
  df <- df %>% select(-"christian")
}

df$christian <- ifelse(df$religion == 3, 1, 0)
df$christian <- as.factor(df$christian)

recoded_vars <- append(recoded_vars, "christian")
```

capture drop low_caste
gen low_caste = .
replace low_caste = 0 if caste == 1
replace low_caste = 1 if caste > 1 & caste < 6


```{r}
# Drop variables starting with 'caste'
if ("low_caste" %in% names(df)) {
  df <- df %>% select(-starts_with("low_caste"))
}

# Replace values based on conditions
df <- df %>% dplyr::mutate(low_caste = ifelse(caste == "1. None", 0,1) %>% as.factor())
df$low_caste <- df$low_caste %>% as.factor()

# df <- df %>% dplyr::mutate(caste = case_when(df$caste == `1. None` ~ 4, 
#                                        `2. Other backward caste` ~ 3, 
#                                        `3. Scheduled Tribe` ~ 2,  
#                                        `4. Scheduled Caste` 1))

recoded_vars <- append(recoded_vars, "low_caste")
```


capture drop toilet
gen toilet = . 
replace toilet = 0 if toilet_type == 0 
replace toilet = 1 if toilet_type  == 1 | toilet_type == 2

```{r}
# Remove 'toilet' column if it exists
if ("toilet" %in% colnames(df)) {
  df$toilet <- NULL
}

# Create 'toilet' column
df$toilet <- NA

# Replace 'toilet' values based on 'toilet_type'
df$toilet[df$toilet_type == "0. no toilet" ] <- 0
df$toilet[df$toilet_type == "1. flush toilet" | df$toilet_type ==  "2. pit or dry toilet"] <- 1

df$toilet <- as.factor(df$toilet)

recoded_vars <- append(recoded_vars, "toilet")
```


gen m_height = round(maternal_height_cm,1)
gen m_ht152 = cond(m_height >= 152 & m_height ~=.,1,0,0)

```{r}
# Create 'm_height' variable and round 'maternal_height_cm' to 1 decimal place
df <- df %>% mutate(m_height = round(maternal_height_cm, 1))

# Create 'm_ht152' variable
df <- df %>% mutate(m_ht152 = if_else(m_height >= 152 & !is.na(m_height), 1, 0))

recoded_vars <- append(recoded_vars, "m_height")
```


capture drop iron_3mo
recode preg_iron (0 1 2 = 0 "0. no")(3 4 = 1 "1. yes"),gen(iron_3mo)

```{r}
# remove iron_3mo if it exists
if ("iron_3mo" %in% names(df)) {
  df <- df %>% select(-iron_3mo)
}

# Recode 'preg_iron' variable into 'iron_3mo' variable
df <- df %>% mutate(iron_3mo = case_when(
  preg_iron %in% c("0", "1. 1 mo or less", "2. 1-3 mo") ~ "0. no",
  preg_iron %in% c("3. 3-6 mo", "4. 6+ mo") ~ "1. yes",
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "iron_3mo")
```


gen maternal_education = mother_education
replace maternal_education = 3 if mother_education == 4

```{r}
# Create 'maternal_education' variable and assign 'mother_education' values
df <- df %>% mutate(maternal_education = mother_education)

# Replace 'maternal_education' value with 3 if 'mother_education' equals 4
df <- df %>% mutate(maternal_education = if_else(mother_education == 4, 3, maternal_education))

# Create label for 'maternal_education' variable
df$maternal_education <- factor(df$maternal_education, levels = c(0, 1, 2, 3),
                                labels = c("no edu", "primary", "secondary", "high secondary or college"))

recoded_vars <- append(recoded_vars, "maternal_education")
```


capture drop anc_4visits
recode anc_visits (0 1 = 0 "0. no")(2 = 1 "1. yes"),gen(anc_4visits)

```{r}
# remove anc_4visits if it exists
if ("anc_4visits" %in% names(df)) {
  df <- df %>% select(-anc_4visits)
}

# Recode 'anc_visits' variable into 'anc_4visits' variable
df <- df %>% mutate(anc_4visits = case_when(
  anc_visits %in% c("0. no visits", "1. 1-3 visits") ~ "0. no",
  anc_visits == "2. 4 or more visits" ~ "1. yes",
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "anc_4visits")
```


capture drop m_ht
gen m_ht = . 
replace m_ht = 1 if maternal_height_cm >= 157.5 & maternal_height_cm !=.
replace m_ht = 2 if maternal_height_cm >= 147.3 & maternal_height_cm < 157.5
replace m_ht = 3 if maternal_height_cm < 147.3
label define m_ht 1 "1 5feet2 or higher" 2 "4feet8 to 5feet2" 3 "less than 4feet8"
label value m_ht m_ht

```{r}
# Drop m_ht column if it exists
if("m_ht" %in% names(df)) {
  df$m_ht <- NULL
}

# Create m_ht column with missing values
df$m_ht <- NA

# Assign values to m_ht based on maternal height
df$m_ht <- ifelse(df$maternal_height_cm >= 157.5 & !is.na(df$maternal_height_cm), 1,
                  ifelse(df$maternal_height_cm >= 147.3 & df$maternal_height_cm < 157.5, 2,
                         ifelse(df$maternal_height_cm < 147.3, 3, df$m_ht)))

# Add labels to m_ht
labels <- c("1. 5feet2 or higher", "2. 4feet8 to 5feet2", "3. less than 4feet8")
levels <- c(1, 2, 3)
df$m_ht <- factor(df$m_ht, levels = levels, labels = labels)

recoded_vars <- append(recoded_vars, "m_ht")
```


gen sibs = cond(siblings >= 3 & siblings ~=.,1,0,0) 
label define house_finish 0  "0. no concrete material" 1  "1. some concrete material" 2  "2. all concrete"
label value house_finish house_finish
label define maternal_education 0  "no edu" 1  "primary" 2  "secondary" 3  "high secondary or college"
label value maternal_education maternal_education

```{r}
num_siblings = c("3. 3", "4. 4 or more")

# Create 'sibs' variable
df <- df %>% mutate(sibs = case_when(
  siblings %in% num_siblings  ~ 1,
  TRUE ~ 0 # all other cases
))

recoded_vars <- append(recoded_vars, "sibs")
```


```{r}
df$house_partially_finished_no <- ifelse(df$house_finish != "1. some concrete material", 1, 0)  %>% as.factor()

recoded_vars <- append(recoded_vars, "house_partially_finished_no")
```


capture drop mother_agecat
recode v013 (1/2 = 1 "1. <25")(3/4 = 2 "25-34")(5/7 = 3 "35+"), gen(mother_agecat)

```{r}
# remove mother_agecat if it exists
if ("mother_agecat" %in% names(df)) {
  df <- df %>% select(-mother_agecat)
}

# Recode 'v013' variable into 'mother_agecat' variable
df <- df %>% mutate(mother_agecat = case_when(
  v013 %in% c(1, 2) ~ "1. <25",
  v013 %in% c(3, 4) ~ "2. 25-34",
  v013 %in% c(5, 6, 7) ~ "3. 35+",
  TRUE ~ NA_character_ # all other cases
))

recoded_vars <- append(recoded_vars, "mother_agecat")
```


capture drop scale_ciaf1_house
gen scale_ciaf1_house = .
replace scale_ciaf1_house = 0 if house_finish == 2
replace scale_ciaf1_house = 3 if house_finish == 1
replace scale_ciaf1_house = 4.5 if house_finish == 0

```{r}
# Drop scale_ciaf1_house column if it exists
if("scale_ciaf1_house" %in% names(df)) {
  df$scale_ciaf1_house <- NULL
}

# Create scale_ciaf1_house column with missing values
df$scale_ciaf1_house <- NA

# Assign values to scale_ciaf1_house based on house_finish
df$scale_ciaf1_house <- ifelse(df$house_finish == "2. all concrete", 0,
                               ifelse(df$house_finish == "1. some concrete material", 3,
                                      ifelse(df$house_finish == "0. no concrete material", 4.5, df$scale_ciaf1_house)))

recoded_vars <- append(recoded_vars, "scale_ciaf1_house")
```


capture drop scale_ciaf1_mheight
gen scale_ciaf1_mheight = . 
replace scale_ciaf1_mheight = 0 if m_ht == 1 
replace scale_ciaf1_mheight = 4 if m_ht == 2
replace scale_ciaf1_mheight = 8 if m_ht == 3

```{r}
# Drop scale_ciaf1_mheight column if it exists
if("scale_ciaf1_mheight" %in% names(df)) {
  df$scale_ciaf1_mheight <- NULL
}

# Create scale_ciaf1_mheight column with missing values
df$scale_ciaf1_mheight <- NA

# Assign values to scale_ciaf1_mheight based on m_ht
df$scale_ciaf1_mheight <- ifelse(df$m_ht == "1. 5feet2 or higher", 0,
                                 ifelse(df$m_ht == "2. 4feet8 to 5feet2", 4,
                                        ifelse(df$m_ht == "3. less than 4feet8", 8, df$scale_ciaf1_mheight)))

recoded_vars <- append(recoded_vars, "scale_ciaf1_mheight")
```


capture drop scale_ciaf1_medu
gen scale_ciaf1_medu = .
replace scale_ciaf1_medu = 0 if maternal_education == 3
replace scale_ciaf1_medu = 2 if maternal_education == 2
replace scale_ciaf1_medu = 3.5 if maternal_education == 1
replace scale_ciaf1_medu = 5 if maternal_education == 0 

```{r}
# Drop scale_ciaf1_medu column if it exists
if("scale_ciaf1_medu" %in% names(df)) {
  df$scale_ciaf1_medu <- NULL
}

# Create scale_ciaf1_medu column with missing values
df$scale_ciaf1_medu <- NA

# Assign values to scale_ciaf1_medu based on maternal_education
df$scale_ciaf1_medu <- ifelse(df$maternal_education == "high secondary or college", 0,
                              ifelse(df$maternal_education == "secondary", 2,
                                     ifelse(df$maternal_education == "primary", 3.5,
                                            ifelse(df$maternal_education == "no edu", 5, df$scale_ciaf1_medu))))

recoded_vars <- append(recoded_vars, "scale_ciaf1_medu")
```


capture drop scale_ciaf1_toilet
gen scale_ciaf1_toilet = . 
replace scale_ciaf1_toilet = 0 if toilet == 1
replace scale_ciaf1_toilet = 3 if toilet == 0 

```{r}
# Drop scale_ciaf1_medu column if it exists
if("scale_ciaf1_toilet" %in% names(df)) {
  df$scale_ciaf1_toilet <- NULL
}

# Create 'scale_ciaf1_toilet' column based on 'toilet' values
df$scale_ciaf1_toilet <- ifelse(df$toilet == 1, 0,
                                 ifelse(df$toilet == 0, 3, NA))

recoded_vars <- append(recoded_vars, "scale_ciaf1_toilet")
```


capture drop scale_ciaf1_zone
gen scale_ciaf1_zone = . 
replace scale_ciaf1_zone = 0 if state_zones == 1
replace scale_ciaf1_zone = 2.5 if state_zones == 2 | state_zones == 3
replace scale_ciaf1_zone = 2 if state_zones == 4 | state_zones == 6
replace scale_ciaf1_zone = 5 if state_zones == 5

```{r}
# Drop scale_ciaf1_zone column if it exists
if("scale_ciaf1_zone" %in% names(df)) {
  df$scale_ciaf1_zone <- NULL
}

# Create scale_ciaf1_zone column with missing values
df$scale_ciaf1_zone <- NA

# Assign values to scale_ciaf1_zone based on state_zones
df$scale_ciaf1_zone <- ifelse(df$state_zones == "1. North-Eastern", 0,
                              ifelse(df$state_zones %in% c("2. North", "3. Central"), 2.5,
                                     ifelse(df$state_zones %in% c("4. Eastern", "6. Southern"), 2,
                                            ifelse(df$state_zones == "5. Western", 5, df$scale_ciaf1_zone))))

recoded_vars <- append(recoded_vars, "scale_ciaf1_zone")
```


capture drop scale_ciaf1_fuel
gen scale_ciaf1_fuel = 0 if cooking_fuel == 1
replace scale_ciaf1_fuel = 0.5 if cooking_fuel == 0 

```{r}
# Drop scale_ciaf1_fuel column if it exists
if("scale_ciaf1_fuel" %in% names(df)) {
  df$scale_ciaf1_fuel <- NULL
}

# Create scale_ciaf1_fuel column with zeros
df$scale_ciaf1_fuel <- 0

# Assign values to scale_ciaf1_fuel based on cooking_fuel
df$scale_ciaf1_fuel <- ifelse(df$cooking_fuel == 1, 0,
                              ifelse(df$cooking_fuel == 0, 0.5, df$scale_ciaf1_fuel))

recoded_vars <- append(recoded_vars, "scale_ciaf1_fuel")
```


capture drop scale_ciaf1_female
gen scale_ciaf1_female = .
replace scale_ciaf1_female = 0 if female == 1 
replace scale_ciaf1_female = 1 if female == 0

```{r}
# Drop scale_ciaf1_female column if it exists
if("scale_ciaf1_female" %in% names(df)) {
  df$scale_ciaf1_female <- NULL
}

# Create scale_ciaf1_female column with missing values
df$scale_ciaf1_female <- NA

# Assign values to scale_ciaf1_female based on female
df$scale_ciaf1_female <- ifelse(df$female == "1. female", 0,
                                ifelse(df$female == "0. male", 1, df$scale_ciaf1_female))

recoded_vars <- append(recoded_vars, "scale_ciaf1_female")
```


capture drop scale_ciaf1_caste
gen scale_ciaf1_caste = . 
replace scale_ciaf1_caste = 0 if low_caste == 0
replace scale_ciaf1_caste = 2.5 if low_caste == 1

```{r}
# Drop scale_ciaf1_caste column if it exists
if("scale_ciaf1_caste" %in% names(df)) {
  df$scale_ciaf1_caste <- NULL
}

# Create scale_ciaf1_caste column with missing values
df$scale_ciaf1_caste <- NA

# Assign values to scale_ciaf1_caste based on low_caste
df$scale_ciaf1_caste <- ifelse(df$low_caste == 0, 0,
                               ifelse(df$low_caste == 1, 2.5, df$scale_ciaf1_caste))

recoded_vars <- append(recoded_vars, "scale_ciaf1_caste")
```


capture drop scale_ciaf1_priority
gen scale_ciaf1_priority = . 
replace scale_ciaf1_priority = 0 if priority_state == 0 
replace scale_ciaf1_priority = 1 if priority_state == 1

```{r}
# Drop scale_ciaf1_priority column if it exists
if("scale_ciaf1_priority" %in% names(df)) {
  df$scale_ciaf1_priority <- NULL
}

# Create scale_ciaf1_priority column with missing values
df$scale_ciaf1_priority <- NA

# Assign values to scale_ciaf1_priority based on priority_state
df$scale_ciaf1_priority <- ifelse(df$priority_state == 0, 0,
                                  ifelse(df$priority_state == 1, 1, df$scale_ciaf1_priority))

recoded_vars <- append(recoded_vars, "scale_ciaf1_priority")
```


capture drop score_ciaf1
egen score_ciaf1 = rowtotal(scale_ciaf1*)

```{r}
# Drop score_ciaf1 column if it exists
if("score_ciaf1" %in% names(df)) {
  df$score_ciaf1 <- NULL
}

# Create score_ciaf1 column by summing up the scale_ciaf1* columns
df$score_ciaf1 <- rowSums(df[grep("scale_ciaf1", names(df))], na.rm = TRUE)

recoded_vars <- append(recoded_vars, "score_ciaf1")
```


capture dist_ciaf_mean
by shdistri, sort : egen float dist_ciaf_mean = mean(ciaf)
capture drop mean_ciaf_dist
recode dist_ciaf_mean (0/0.399999 = 1 "1. < 40%")(0.40/0.6 = 2 "2. 40-60%")(0.6000001/1.0 = 3 "3. > 60%"),gen(mean_ciaf_dist)
gen dist_ciaf_decile = round(dist_ciaf_mean,0.1)
capture drop ciaf_decile
recode dist_ciaf_decile (0.2 .3 = 1 "< 30%")(0.4 = 2 "40%") (0.5 = 3 "50%")(0.6 = 4 "60%")(0.7 = 5 "70%")(0.8 = 6 "80%"),

```{r}
# shdistri was renamed to shdist in the latest DHS

# Calculate mean CIAF by shdist and sort
df <- df %>% group_by(shdist) %>% arrange(shdist) %>% 
  mutate(dist_ciaf_mean = mean(ciaf, na.rm = TRUE)) %>% ungroup()

# Drop mean_ciaf_dist column if it exists
if("mean_ciaf_dist" %in% names(df)) {
  df$mean_ciaf_dist <- NULL
}

# Recode dist_ciaf_mean into mean_ciaf_dist based on cutoffs
df$mean_ciaf_dist <- cut(df$dist_ciaf_mean, 
                         breaks = c(0, 0.4, 0.6, 1.0), 
                         labels = c("1. < 40%", "2. 40-60%", "3. > 60%"))

# Round dist_ciaf_mean to one decimal place and assign to dist_ciaf_decile
df$dist_ciaf_decile <- round(df$dist_ciaf_mean, 1)

# Drop ciaf_decile column if it exists
if("ciaf_decile" %in% names(df)) {
  df$ciaf_decile <- NULL
}

# Recode dist_ciaf_decile into ciaf_decile based on cutoffs
df$ciaf_decile <- cut(df$dist_ciaf_decile, 
                      breaks = c(-Inf, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, Inf), 
                      labels = c("< 30%", "40%", "50%", "60%", "70%", "80%", "> 80%"))

recoded_vars <- append(recoded_vars, "dist_ciaf_mean")
recoded_vars <- append(recoded_vars, "mean_ciaf_dist")
recoded_vars <- append(recoded_vars, "dist_ciaf_decile")
recoded_vars <- append(recoded_vars, "ciaf_decile")
```

```{r}
df$mean_district_CIAF_40_49.9 <- ifelse(df$ciaf_decile == "40%", 1, 0)  %>% as.factor()
df$mean_district_CIAF_50_59.9 <- ifelse(df$ciaf_decile == "50%", 1, 0)  %>% as.factor()
df$mean_district_CIAF_60_69.9 <- ifelse(df$ciaf_decile == "60%", 1, 0)  %>% as.factor()
df$mean_district_CIAF_70_79.9 <- ifelse(df$ciaf_decile == "70%", 1, 0) %>% as.factor()
df$mean_district_CIAF_80_89.9 <- ifelse(df$ciaf_decile == "80%", 1, 0) %>% as.factor()

recoded_vars <- append(recoded_vars, "mean_district_CIAF_40_49.9")
recoded_vars <- append(recoded_vars, "mean_district_CIAF_50_59.9")
recoded_vars <- append(recoded_vars, "mean_district_CIAF_60_69.9")
recoded_vars <- append(recoded_vars, "mean_district_CIAF_70_79.9")
recoded_vars <- append(recoded_vars, "mean_district_CIAF_80_89.9")
```

```{r}
# Convert CIAF to factor after using is as numeric before
df$ciaf <- as.factor(df$ciaf)
```


capture dist_wasted_mean
by shdistri, sort : egen float dist_wasted_mean = mean(wasted)
capture drop mean_wasted_dist
recode dist_wasted_mean (0/0.399999 = 1 "1. < 40%")(0.40/0.6 = 2 "2. 40-60%")(0.6000001/1.0 = 3 "3. > 60%"),gen(mean_wasted_dist)
gen dist_wasted_decile = round(dist_wasted_mean,0.1)
capture drop wasted_decile
recode dist_wasted_decile (0 = 0 "0-10%")(0.1 = 1 "10-20%")(.2 = 2 "20-30%")(0.3 = 3 "30-40%") (0.4 = 4 "40-50%")(0.5 = 5 "50-60%") ///
, gen(wasted_decile)

```{r}
df <- df %>% 
  mutate(wasted = as.integer(wasted)) %>%
  mutate(wasted = case_when(wasted == 1 ~ 0,
                            wasted == 2 ~ 1))

# Calculate the mean 'wasted' by 'shdist' and create a new column 'dist_wasted_mean'
df <- df %>%
  group_by(shdist) %>%
  mutate(dist_wasted_mean = mean(wasted, na.rm = TRUE)) %>%
  ungroup()

# Create a new column 'mean_wasted_dist' based on the conditions specified
df$mean_wasted_dist <- cut(df$dist_wasted_mean,
                           breaks = c(-Inf, 0.4, 0.6, Inf),
                           labels = c("1. < 40%", "2. 40-60%", "3. > 60%"),
                           right = FALSE)

# Create a new column 'dist_wasted_decile' by rounding 'dist_wasted_mean' to the nearest tenth
df$dist_wasted_decile <- round(df$dist_wasted_mean, 1)

# Create a new column 'wasted_decile' based on the conditions specified
df$wasted_decile <- cut(df$dist_wasted_decile,
                        breaks = c(0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6),
                        labels = c("0-10%", "10-20%", "20-30%", "30-40%", "40-50%", "50-60%"),
                        right = FALSE)

recoded_vars <- append(recoded_vars, "mean_wasted_dist")
recoded_vars <- append(recoded_vars, "dist_wasted_decile")
recoded_vars <- append(recoded_vars, "wasted_decile")
```


capture drop priority_district 
capture drop dist_priority
recode shdist (300 301 305 313 317 318 204  205 206 209 210 211 212 214 236 238 401 402 406 416 417 351 352 353 356 358 359 362 366 367 368 369 424 425 426 427 428 429 431 441 446 453 454 460 461 462 463 464 465 389 390 391 394 396 397 398 399 106 107 114 115 116 121 123 124 125 130 149 150 151 152 153 154 155 174 176 177 180 181 182 183 186 200 201 202 59 61 68 543 544 551 552 245 247 249 252 255 256 258 90 92 468 469 472 484 485 487 489 491 75 77 80 87 89 23 25 27 34 3 5 6 16 17 18 556 557 559 560 561 565 579 580 588 592 593 272 273 274 280 293 295 296 299 281 286 287 288 634 35 39 44 47 54 242 605 606 614 623 628 631 291 329 330 332 333 343 261 262 267 269 270 271 532 538 511 523 499 498 497 508 514 515 512 279 625 184 = 9999) , gen(dist_priority)
replace dist_priority = 0 if dist_priority != 9999
replace dist_priority = 1 if dist_priority == 9999

```{r}
# Create or modify the 'dist_priority' variable
df <- df %>%
  mutate(
    dist_priority = if_else(
      shdist %in% c(
        300, 301, 305, 313, 317, 318, 204, 205, 206, 209, 210, 211, 212, 214, 236, 238,
        401, 402, 406, 416, 417, 351, 352, 353, 356, 358, 359, 362, 366, 367, 368, 369,
        424, 425, 426, 427, 428, 429, 431, 441, 446, 453, 454, 460, 461, 462, 463, 464,
        465, 389, 390, 391, 394, 396, 397, 398, 399, 106, 107, 114, 115, 116, 121, 123,
        124, 125, 130, 149, 150, 151, 152, 153, 154, 155, 174, 176, 177, 180, 181, 182,
        183, 186, 200, 201, 202, 59, 61, 68, 543, 544, 551, 552, 245, 247, 249, 252,
        255, 256, 258, 90, 92, 468, 469, 472, 484, 485, 487, 489, 491, 75, 77, 80, 87,
        89, 23, 25, 27, 34, 3, 5, 6, 16, 17, 18, 556, 557, 559, 560, 561, 565, 579, 580,
        588, 592, 593, 272, 273, 274, 280, 293, 295, 296, 299, 281, 286, 287, 288, 634,
        35, 39, 44, 47, 54, 242, 605, 606, 614, 623, 628, 631, 291, 329, 330, 332, 333,
        343, 261, 262, 267, 269, 270, 271, 532, 538, 511, 523, 499, 498, 497, 508, 514,
        515, 512, 279, 625, 184
      ),
      1,  # Assign 1 for matching values
      0   # Assign 0 for non-matching values
    )
  )


```



```{r}
rev(colnames(df))
```


## Exclusion/Inclusion Criteria 
```{r}

df <- df %>% dplyr::filter(child_age < 60)

print(paste0("Filter only children under the age of 5 (60 months): ", nrow(df)))

df <- df %>% dplyr::filter(defacto_child == 1)

print(paste0("Children that are part of the household: ", nrow(df)))

df <- df %>% dplyr::filter(child_alive == 1)

print(paste0("Children that are alive: ", nrow(df)))

df <- df %>% dplyr::filter(!is.na(ciaf))

print(paste0("Children with complete anthropometric information: ", nrow(df)))

```




```{r}
# Save recoded data as R object
dffile <- here(data_folder, "recoded.rda") 
save(df, file=dffile)
```

```{r}
# Save list of recoded variables
rvfile <- here(data_folder, "recoded-variables.txt")
write.table(recoded_vars, rvfile, row.names=FALSE, col.names=FALSE)
```
